### 2020-03-01 #2133 @whedon
[PRE REVIEW]: VlaPy: A Python package for Eulerian Vlasov-Poisson-Fokker-Planck Simulations
**Submitting author:** @joglekara (<a href="http://orcid.org/0000-0003-3599-5629">Archis Joglekar</a>)
**Repository:** <a href="https://github.com/joglekara/VlaPy" target ="_blank">https://github.com/joglekara/VlaPy</a>
**Version:** v1.0
