### 2020-03-01 #2133 @joglekara
👋 all!

I am honored and excited to contribute.

Here are some potential reviewers I found from the Google Sheet:
TomGoffrey
StanczakDominik
ali-ramadhan

On Sun, Mar 1, 2020 at 5:01 PM Lorena A. Barba <notifications@github.com>
wrote:

> 👋 @katyhuff <https://github.com/katyhuff> — Could you handle this
> submission as editor? It is a pure-Python solver for the
> Vlasov-Poisson-Fokker-Planck based on finite-difference and pseudo-spectral
> methods. Thanks!
>
