### 2020-03-10 #2133 @dpsanders
👋 Hello @joglekara. Thanks for your submission to JOSS.

Looking over the repo, I notice that there is no (narrative) documentation. Could you please add documentation on the system being solved (as in the paper) and on how to use your implementation (API) before we proceed to review?

I didn't quite understand the relation between this submission and PlasmaPy.

There's also a missing reference at the end of page 3 in the paper.

Thanks!
