### 2020-03-10 #2133 @arfon
@dpsanders @joglekara - I made a couple of minor readability improvements in https://github.com/joglekara/VlaPy/pull/6
