### 2020-03-13 f8aed75
Manuscript updates - citations - text
