+++ paper/paper.bib
@@ -1,7 +1,6 @@
 @article{Banks2017,
 author = {Banks, J W and Brunner, S and Berger, R L and Arrighi, W J and Tran, T M},
 doi = {10.1103/PhysRevE.96.043208},
-file = {:Users/archis/Dropbox/work/Papers/Mac2/2017/Banks et al. - Physical Review E - 2017.pdf:pdf},
 issn = {2470-0045},
 journal = {Physical Review E},
 month = {oct},
@@ -16,7 +15,6 @@ year = {2017}
 @article{Joglekar2018,
 author = {Joglekar, Archis S and Winjum, Benjamin J and Tableman, Adam and Wen, Han and Tzoufras, Michail and Mori, Warren B},
 doi = {10.1088/1361-6587/aab978},
-file = {:Users/archis/Dropbox/work/Papers/Mac2/2018/Joglekar et al. - Plasma Physics and Controlled Fusion - 2018.pdf:pdf},
 issn = {0741-3335},
 journal = {Plasma Physics and Controlled Fusion},
 keywords = {and length-scales as well,appear in colour only,as plasma densities and,cartesian tensor,fokker,in the online journal,kinetic modeling,occur over disparate time-,planck,plasma physics phenomenon can,some fi gures may,spherical harmonic,tempera-,vlasov},
@@ -34,7 +32,6 @@ year = {2018}
 abstract = {A model of plasma oscillations in the presence of small-angle collisions is presented which admits of exact analytic solution. Certain features of the true collision terms are preserved. Namely, the effect of collisions is represented by a diffusion in velocity space, which makes the distribution function tend to the Maxwell distribution, and which conserves the number of particles. In the limit of infrequent collisions the results of Landau are recovered.},
 author = {Lenard, A. and Bernstein, Ira B.},
 doi = {10.1103/PhysRev.112.1456},
-file = {:Users/archis/Dropbox/work/Papers/Mac2/1958/Lenard, Bernstein - Physical Review - 1958.pdf:pdf},
 issn = {0031899X},
 journal = {Physical Review},
 number = {5},
@@ -47,10 +44,8 @@ year = {1958}
 @article{Canosa1973,
 author = {Canosa, Jos{\'{e}}},
 doi = {10.1016/0021-9991(73)90131-9},
-file = {:Users/archis/Dropbox/work/Papers/Mac2/1973/Canosa - Journal of Computational Physics - 1973.pdf:pdf},
 issn = {00219991},
 journal = {Journal of Computational Physics},
-mendeley-groups = {LinearLD+coll},
 month = {sep},
 number = {1},
 pages = {158--160},
@@ -64,10 +59,8 @@ year = {1973}
 abstract = {A brief summary is given of the early studies of Landau damping, followed by a discussion of the issues of singularities in the distribution function, reversibility and nonlinear constraints. A difference is emphasized between the evolution of a single-scale localized Langmuir perturbation and a long quasimonochromatic wavetrain. Applicability conditions of the quasilinear approximation are discussed. Examples of the use of the concept of Landau damping in hydrodynamics, astrophysics and other systems are presented.},
 author = {Ryutov, D D},
 doi = {10.1088/0741-3335/41/3A/001},
-file = {:Users/archis/Dropbox/work/Papers/Mac2/1999/Ryutov - Plasma Physics and Controlled Fusion - 1999.pdf:pdf},
 issn = {0741-3335},
 journal = {Plasma Physics and Controlled Fusion},
-mendeley-groups = {LinearLD+coll},
 number = {3A},
 pages = {A1--A12},
[…]
