### 2020-03-13 #2133 @joglekara
Hi @dpsanders , thanks for your comments. @arfon , thanks for the improvements.

Re: Docs and API:
The API instructions are provided through a jupyter-notebook but we understand that it should be more obvious. We will add some more documentation about the equations being solved along with instructions on how to use the API.

Re: PlasmaPy
This was partially motivated by the explicit mention by the PlasmaPy community of the need for this type of simulation code. There is no greater relationship than that at the present time. We intend to reach out to the PlasmaPy community to collect their feedback as this submission matures.

Re: Missing Reference
Good catch. This has been addressed in an on-going PR.

Re: 2nd Author
Matthew has helped provide high level guidance for the software and has helped write the manuscript.
