### 2020-03-13 4a6cae3
incorporating @arfon's suggested changes
