+++ docs/electricfield.rst
@@ -1,0 +1,10 @@
+Electric Field
+--------------
+
+The Poisson equation is solved pseudospectrally.
+
+This solver is tested to reproduce analytical solutions to a periodic Poisson system.
+
+
+
+...more coming soon...
