+++ docs/landau.rst
@@ -1,0 +1,16 @@
+Landau Damping
+---------------
+
+One of the most fundamental plasma physics phenomenon is that described by Landau damping.
+
+Plasmas can support electrostatic oscillations. The oscillation frequency is given by the electrostatic electron
+plasma wave (EPW) dispersion relation. When a wave of sufficiently small amplitude is driven at the resonant
+wave-number and frequency pairing, there is a resonant exchange of energy between the plasma and the electric field,
+and the electrons can damp the electric field.
+
+In VlaPy, we verify that the damping rate is reproduced for a few different wave numbers.
+This is shown in `notebooks/landau_damping.ipynb.`
+
+We include validation against this phenomenon as an integrated test.
+
+...this page is in development...
