### 2020-04-05 ede9b62
a little bit of refactoring
