+++ docs/bibs/other.bib
@@ -1,0 +1,24 @@
+@article{Hoyer2017,
+abstract = {xarray is an open source project and Python package that provides a toolkit and data structures for N-dimensional labeled arrays. Our approach combines an application programing interface (API) inspired by pandas with the Common Data Model for self-described scientific data. Key features of the xarray package include label-based indexing and arithmetic, interoperability with the core scientific Python packages (e.g., pandas, NumPy, Matplotlib), out-of-core computation on datasets that don't fit into memory, a wide range of serialization and input/output (I/O) options, and advanced multi-dimensional data manipulation tools such as group-by and resampling. xarray, as a data model and analytics toolkit, has been widely adopted in the geoscience community but is also used more broadly for multi-dimensional data analysis in physics, machine learning and finance.},
+author = {Hoyer, Stephan and Hamman, Joseph J.},
+doi = {10.5334/jors.148},
+file = {:Users/archis/Dropbox/work/Papers/Mac2/2017/Hoyer, Hamman - Journal of Open Research Software - 2017.pdf:pdf},
+issn = {2049-9647},
+journal = {Journal of Open Research Software},
+keywords = {data,data analysis,data handling,multidimensional,netcdf,pandas,python},
+month = {apr},
+pages = {1--6},
+title = {{xarray: N-D labeled Arrays and Datasets in Python}},
+url = {http://openresearchsoftware.metajnl.com/articles/10.5334/jors.148/},
+volume = {5},
+year = {2017}
+}
+@article{Zaharia2018,
+  title={Accelerating the Machine Learning Lifecycle with MLflow.},
+  author={Zaharia, Matei and Chen, Andrew and Davidson, Aaron and Ghodsi, Ali and Hong, Sue Ann and Konwinski, Andy and Murching, Siddharth and Nykodym, Tomas and Ogilvie, Paul and Parkhe, Mani and others},
+  journal={IEEE Data Eng. Bull.},
+  volume={41},
+  number={4},
+  pages={39--45},
+  year={2018}
+}
