+++ docs/bibs/vlasov.bib
@@ -1,0 +1,62 @@
+@article{Omelyan2002,
+abstract = {An approach is proposed to improve the efficiency of fourth-order algorithms for numerical integration of the equations of motion in molecular dynamics simulations. The approach is based on an extension of the decomposition scheme by introducing extra evolution subpropagators. The extended set of parameters of the integration is then determined by reducing the norm of truncation terms to a minimum. In such a way, we derive new explicit symplectic Forest-Ruth- and Suzuki-like integrators and present them in time-reversible velocity and position forms. It is proven that these optimized integrators lead to the best accuracy in the calculations at the same computational cost among all possible algorithms of the fourth order from a given decomposition class. It is shown also that the Forest-Ruth-like algorithms, which are based on direct decomposition of exponential propagators, provide better optimization than their Suzuki-like counterparts which represent compositions of second-order schemes. In particular, using our optimized Forest-Ruth-like algorithms allows us to increase the efficiency of the computations by more than ten times with respect to that of the original integrator by Forest and Ruth, and by approximately five times with respect to Suzuki's approach. The theoretical predictions are confirmed in molecular dynamics simulations of a Lennard-Jones fluid. A special case of the optimization of the proposed Forest-Ruth-like algorithms to celestial mechanics simulations is considered as well. {\textcopyright} 2002 Elsevier Science B.V. All rights reserved.},
+archivePrefix = {arXiv},
+arxivId = {cond-mat/0110585},
+author = {Omelyan, I.P. and Mryglod, I.M. and Folk, R.},
+doi = {10.1016/S0010-4655(02)00451-4},
+eprint = {0110585},
+file = {:Users/archis/Dropbox/work/Papers/Mac2/2002/Omelyan, Mryglod, Folk - Computer Physics Communications - 2002.pdf:pdf},
+issn = {00104655},
+journal = {Computer Physics Communications},
+month = {jul},
+number = {2},
+pages = {188--202},
+primaryClass = {cond-mat},
+title = {{Optimized Forest–Ruth- and Suzuki-like algorithms for integration of motion in many-body systems}},
+url = {https://linkinghub.elsevier.com/retrieve/pii/S0010465502004514},
+volume = {146},
+year = {2002}
+}
+@article{Ryutov1999,
+abstract = {A brief summary is given of the early studies of Landau damping, followed by a discussion of the issues of singularities in the distribution function, reversibility and nonlinear constraints. A difference is emphasized between the evolution of a single-scale localized Langmuir perturbation and a long quasimonochromatic wavetrain. Applicability conditions of the quasilinear approximation are discussed. Examples of the use of the concept of Landau damping in hydrodynamics, astrophysics and other systems are presented.},
+author = {Ryutov, D D},
+doi = {10.1088/0741-3335/41/3A/001},
+file = {:Users/archis/Dropbox/work/Papers/Mac2/1999/Ryutov - Plasma Physics and Controlled Fusion - 1999.pdf:pdf},
+issn = {0741-3335},
+journal = {Plasma Physics and Controlled Fusion},
+mendeley-groups = {LinearLD+coll},
+number = {3A},
+pages = {A1--A12},
+title = {{Landau damping: half a century with the great discovery}},
+volume = {41},
+year = {1999}
+}
+@article{Strang1968,
+author = {Strang, Gilbert},
+doi = {10.1137/0705041},
+file = {:Users/archis/Dropbox/work/Papers/Mac2/1968/Strang - SIAM Journal on Numerical Analysis - 1968.pdf:pdf},
+issn = {0036-1429},
+journal = {SIAM Journal on Numerical Analysis},
+month = {sep},
+number = {3},
+pages = {506--517},
+title = {{On the Construction and Comparison of Difference Schemes}},
[…]
