+++ docs/bibs/fokkerplanck.bib
@@ -2,7 +2,6 @@
 abstract = {A model of plasma oscillations in the presence of small-angle collisions is presented which admits of exact analytic solution. Certain features of the true collision terms are preserved. Namely, the effect of collisions is represented by a diffusion in velocity space, which makes the distribution function tend to the Maxwell distribution, and which conserves the number of particles. In the limit of infrequent collisions the results of Landau are recovered.},
 author = {Lenard, A. and Bernstein, Ira B.},
 doi = {10.1103/PhysRev.112.1456},
-file = {:Users/archis/Dropbox/work/Papers/Mac2/1958/Lenard, Bernstein - Physical Review - 1958.pdf:pdf},
 issn = {0031899X},
 journal = {Physical Review},
 number = {5},
