+++ docs/bibs/vlasov.bib
@@ -5,7 +5,6 @@ arxivId = {cond-mat/0110585},
 author = {Omelyan, I.P. and Mryglod, I.M. and Folk, R.},
 doi = {10.1016/S0010-4655(02)00451-4},
 eprint = {0110585},
-file = {:Users/archis/Dropbox/work/Papers/Mac2/2002/Omelyan, Mryglod, Folk - Computer Physics Communications - 2002.pdf:pdf},
 issn = {00104655},
 journal = {Computer Physics Communications},
 month = {jul},
@@ -13,7 +12,6 @@ number = {2},
 pages = {188--202},
 primaryClass = {cond-mat},
 title = {{Optimized Forest–Ruth- and Suzuki-like algorithms for integration of motion in many-body systems}},
-url = {https://linkinghub.elsevier.com/retrieve/pii/S0010465502004514},
 volume = {146},
 year = {2002}
 }
@@ -21,7 +19,6 @@ year = {2002}
 abstract = {A brief summary is given of the early studies of Landau damping, followed by a discussion of the issues of singularities in the distribution function, reversibility and nonlinear constraints. A difference is emphasized between the evolution of a single-scale localized Langmuir perturbation and a long quasimonochromatic wavetrain. Applicability conditions of the quasilinear approximation are discussed. Examples of the use of the concept of Landau damping in hydrodynamics, astrophysics and other systems are presented.},
 author = {Ryutov, D D},
 doi = {10.1088/0741-3335/41/3A/001},
-file = {:Users/archis/Dropbox/work/Papers/Mac2/1999/Ryutov - Plasma Physics and Controlled Fusion - 1999.pdf:pdf},
 issn = {0741-3335},
 journal = {Plasma Physics and Controlled Fusion},
 mendeley-groups = {LinearLD+coll},
@@ -34,14 +31,12 @@ year = {1999}
 @article{Strang1968,
 author = {Strang, Gilbert},
 doi = {10.1137/0705041},
-file = {:Users/archis/Dropbox/work/Papers/Mac2/1968/Strang - SIAM Journal on Numerical Analysis - 1968.pdf:pdf},
 issn = {0036-1429},
 journal = {SIAM Journal on Numerical Analysis},
 month = {sep},
 number = {3},
 pages = {506--517},
 title = {{On the Construction and Comparison of Difference Schemes}},
-url = {http://epubs.siam.org/doi/10.1137/0705041},
 volume = {5},
 year = {1968}
 }
@@ -49,14 +44,12 @@ year = {1968}
 abstract = {A convenient, fast, and accurate method of solving the one-dimensional Vlasov equation numerically in configuration space is described. It treats the convective terms in the x and v directions separately and produces a scheme of second order in $\Delta$t. The resulting free-streaming and accelerating equations are computed with Fourier interpolation and spline interpolation methods respectively. The numerical method is tested witth linear and nonlinear problems. The method is very accurate and efficient. A new method of smoothing the distribution function is given. It reduces the computational effort by artificially increasing the entropy of the system. As a result, the distribution function is smooth enough to be well represented on a given mesh. The methods can be generalized in a straightforward way to deal with more complicated cases such as problems with nonperiodic spatial boundary conditions, two- and three-dimensional problems with and without external magnetic and/or electric fields. {\textcopyright} 1976.},
 author = {Cheng, C.Z and Knorr, Georg},
 doi = {10.1016/0021-9991(76)90053-X},
-file = {:Users/archis/Dropbox/work/Papers/Mac2/1976/Cheng, Knorr - Journal of Computational Physics - 1976.pdf:pdf},
 issn = {00219991},
 journal = {Journal of Computational Physics},
 month = {nov},
[…]
