+++ docs/code.rst
@@ -1,0 +1,10 @@
+Code
+---------------
+
+.. toctree::
+   :maxdepth: 2
+
+   code/vlasov
+   code/fokker-planck
+   code/electricfield
+   code/other
