### 2020-04-09 9a6c5db
storing the electric field driver to file too. this will be an essential diagnostic
