+++ vlapy/storage.py
@@ -7,6 +7,9 @@ class StorageManager:
     def __init__(self, xax, vax, tax, base_path):
         self.base_path = base_path
         self.efield_path = os.path.join(base_path, "electric_field_vs_time.nc")
+        self.driver_efield_path = os.path.join(
+            base_path, "driver_electric_field_vs_time.nc"
+        )
         self.f_path = os.path.join(base_path, "dist_func_vs_time.nc")
 
         self.__init_electric_field_storage(tax=tax, xax=xax)
@@ -14,6 +17,7 @@ class StorageManager:
 
     def close(self):
         self.efield_arr.to_netcdf(self.efield_path)
+        self.driver_efield_arr.to_netcdf(self.driver_efield_path)
         self.f_arr.to_netcdf(self.f_path)
 
     def __init_electric_field_storage(self, tax, xax):
@@ -36,6 +40,16 @@ class StorageManager:
 
         self.efield_arr = xr.open_dataarray(self.efield_path)
 
+        driver_electric_field_store = np.zeros((tax.size, xax.size))
+
+        driver_ef_DA = xr.DataArray(
+            data=driver_electric_field_store, coords=[("time", tax), ("space", xax)]
+        )
+
+        driver_ef_DA.to_netcdf(self.driver_efield_path)
+
+        self.driver_efield_arr = xr.open_dataarray(self.driver_efield_path)
+
     def __init_dist_func_storage(self, tax, xax, vax):
         """
         Initialize distribution function storage
@@ -57,7 +71,7 @@ class StorageManager:
 
         self.f_arr = xr.open_dataarray(self.f_path)
 
-    def batched_write_to_file(self, t_range, e, f):
+    def batched_write_to_file(self, t_range, e, e_driver, f):
         """
         Write batched to file
 
@@ -67,17 +81,21 @@ class StorageManager:
 
         :param t_range:
         :param e:
+        :param e_driver:
         :param f:
         :return:
         """
         t_xr = xr.DataArray(data=t_range, dims=["time"])
 
         self.efield_arr.loc[t_xr, :] = e
+        self.driver_efield_arr.loc[t_xr, :] = e_driver
         self.f_arr.loc[t_xr, :] = f
 
         # Save and reopen
         self.efield_arr.to_netcdf(self.efield_path)
+        self.driver_efield_arr.to_netcdf(self.driver_efield_path)
         self.f_arr.to_netcdf(self.f_path)
 
         self.efield_arr = xr.open_dataarray(self.efield_path)
+        self.driver_efield_arr = xr.open_dataarray(self.driver_efield_path)
         self.f_arr = xr.open_dataarray(self.f_path)
