### 2020-04-09 3266963
Replaced numba accelerated routines with broadcasting. Barely faster, but now its more native to numpy, which is the ultimate goal
