+++ vlapy/core/vlasov.py
@@ -1,5 +1,4 @@
 import numpy as np
-from numba import njit
 
 
 def __get_k__(ax):
@@ -40,21 +39,25 @@ def update_velocity_adv_spectral(f, kv, e, dt):
     return np.real(np.fft.ifft(__edfdv__(np.fft.fft(f, axis=1), e, kv, dt), axis=1))
 
 
-@njit
 def __edfdv__(fp, e, kv, dt):
-    for ix in range(e.size):
-        fp[ix,] = (
-            np.exp(-1j * kv * dt * e[ix]) * fp[ix,]
-        )
+    """
 
-    return fp
+    :param fp:
+    :param e:
+    :param kv:
+    :param dt:
+    :return:
+    """
+    return np.exp(-1j * kv * dt * e[:, None]) * fp
 
 
-@njit
 def __vdfdx__(fp, v, kx, dt):
-    for ix in range(kx.size):
-        fp[ix,] = (
-            np.exp(-1j * kx[ix] * dt * v) * fp[ix,]
-        )
+    """
 
-    return fp
+    :param fp:
+    :param v:
+    :param kx:
+    :param dt:
+    :return:
+    """
+    return np.exp(-1j * kx[:, None] * dt * v) * fp
