### 2020-04-09 6313719
reorg wave driver and parameter storage
Now, you can choose which parameters are logged to file and which are logged to the mlflow DB for easy sorting, viz etc. These can be chosen from the diagnostics routines and classes.

The wave driver is now a dictionary that can contain an arbitrary number of pulses at arbitrary (a0,w,k) with arbitrary(-ish) flat-top time envelopes.
