+++ vlapy/manager.py
@@ -8,9 +8,74 @@ from vlapy.core import step, field, lenard_bernstein
 from vlapy import storage
 
 
-def start_run(
-    nx, nv, nt, tmax, nu, w0, k0, a0, diagnostics, name="test", mlflow_path=None
-):
+def _get_rise_fall_coeff_(normalized_time):
+    """
+    This is a smooth function that goes from 0 to 1. It is a 5th order polynomial because it satisfies the following
+    5 conditions
+
+    value = 0 at t=0 and t=1 (2 conditions)
+    first derivative = 0 at t=0 and t=1 (2 conditions)
+    value = 0.5 in the middle
+
+    :param normalized_time:
+    :return:
+    """
+    coeff = (
+        6.0 * pow(normalized_time, 5.0)
+        - 15.0 * pow(normalized_time, 4.0)
+        + 10.0 * pow(normalized_time, 3.0)
+    )
+    return coeff
+
+
+def get_pulse_coefficient(pulse_profile_dictionary, tt):
+    """
+    This function generates an envelope that smoothly goes from 0 to 1, and back down to 0.
+    It follows the nomenclature introduced to me by working with oscilloscopes.
+
+    The pulse profile dictionary will contain a rise time, flat time, and fall time.
+    The rise time is the time over which the pulse goes from 0 to 1
+    The flat time is the duration of the flat-top of the pulse
+    The fall time is the time over which the pulse goes from 1 to 0.
+
+    :param pulse_profile_dictionary:
+    :param tt:
+    :return:
+    """
+    start_time = pulse_profile_dictionary["start_time"]
+    end_time = (
+        start_time
+        + pulse_profile_dictionary["rise_time"]
+        + pulse_profile_dictionary["flat_time"]
+        + pulse_profile_dictionary["fall_time"]
+    )
+
+    this_pulse = 0
+
+    if (tt > start_time) and (tt < end_time):
+        rise_time = pulse_profile_dictionary["rise_time"]
+        flat_time = pulse_profile_dictionary["flat_time"]
+        fall_time = pulse_profile_dictionary["fall_time"]
+        end_rise = start_time + rise_time
+        end_flat = start_time + rise_time + flat_time
+
+        if tt <= end_rise:
+            normalized_time = (tt - start_time) / rise_time
+            this_pulse = _get_rise_fall_coeff_(normalized_time)
+
+        elif (tt > end_rise) and (tt < end_flat):
+            this_pulse = 1.0
+
+        elif tt >= end_flat:
+            normalized_time = (tt - end_flat) / fall_time
+            this_pulse = 1.0 - _get_rise_fall_coeff_(normalized_time)
+
+        this_pulse *= pulse_profile_dictionary["a0"]
+
+    return this_pulse
+
+
+def start_run(all_params, pulse_dictionary, diagnostics, name="test", mlflow_path=None):
     """
     End to end mlflow and xarray storage!!
 
@@ -20,10 +85,11 @@ def start_run(
     :param nv:
     :param nt:
     :param tmax:
-    :param w0:
-    :param k0:
-    :param a0:
+    :param nu:
+    :param pulse_dictionary:
+    :param diagnostics:
     :param name:
+    :param mlflow_path:
     :return:
     """
     if mlflow_path is None:
@@ -34,34 +100,36 @@ def start_run(
     mlflow.set_experiment(name)
 
     with mlflow.start_run():
-        # Log initial conditions
-        params_dict = {
-            "nx": nx,
-            "nv": nv,
-            "w0": w0,
-            "k0": k0,
-            "nt": nt,
-            "tmax": tmax,
-            "a0": a0,
-            "nu": nu,
-        }
[…]
