### 2020-04-09 64b00b3
added a conditional to prevent accumulation of round-off. best practice?
