### 2020-04-12 #2133 @joglekara
Hi @dpsanders , hope you are well at this time.

I've added documentation via ReadTheDocs as well as a quickstart guide to the README.md in the repo.

Looking forward to your feedback.
