### 2020-04-16 b4fdaa2
added a test for it too. this could be made to be standalone but its okay for now
