+++ tests/extras/canosa_values.py
@@ -1,0 +1,119 @@
+k0 = [
+    0.25,
+    0.3,
+    0.35,
+    0.4,
+    0.45,
+    0.5,
+    0.55,
+    0.6,
+    0.65,
+    0.7,
+    0.75,
+    0.8,
+    0.85,
+    0.9,
+    0.95,
+    # 1.0,
+    # 1.05,
+    # 1.1,
+    # 1.15,
+    # 1.2,
+    # 1.25,
+    # 1.3,
+    # 1.35,
+    # 1.4,
+    # 1.45,
+    # 1.5,
+    # 1.55,
+    # 1.6,
+    # 1.65,
+    # 1.7,
+    # 1.75,
+    # 1.8,
+    # 1.85,
+    # 1.9,
+    # 1.95,
+    # 2.0,
+    # 2.3088,
+]
+
+w_real = [
+    1.1056,
+    1.1598,
+    1.2209,
+    1.2850,
+    1.3502,
+    1.4156,
+    1.4809,
+    1.5457,
+    1.6100,
+    1.6739,
+    1.7371,
+    1.7999,
+    1.8621,
+    1.9239,
+    1.9851,
+    # 2.0459,
+    # 2.1062,
+    # 2.1662,
+    # 2.2257,
+    # 2.2848,
+    # 2.3436,
+    # 2.4020,
+    # 2.4600,
+    # 2.5178,
+    # 2.5752,
+    # 2.6323,
+    # 2.6892,
+    # 2.7457,
+    # 2.8020,
+    # 2.8580,
+    # 2.9138,
+    # 2.9693,
+    # 3.0246,
+    # 3.0797,
+    # 3.1345,
+    # 3.1891,
+    # 3.5222,
+]
+
+w_imag = [
+    -0.0021693,
+    -0.012623,
+    -0.034324,
+    -0.066133,
+    -0.10629,
+    -0.15336,
+    -0.20624,
+    -0.26411,
+    -0.32633,
+    -0.39240,
+    -0.46192,
+    -0.53455,
+    -0.61003,
+    -0.68811,
+    -0.76860,
+    # -0.85134,
+    # -0.93615,
+    # -1.0229,
+    # -1.1115,
+    # -1.2019,
+    # -1.2939,
+    # -1.3874,
+    # -1.4824,
+    # -1.5789,
+    # -1.6766,
+    # -1.7757,
+    # -1.8760,
+    # -1.9775,
+    # -2.0801,
+    # -2.1839,
+    # -2.2886,
+    # -2.3944,
+    # -2.5012,
+    # -2.6090,
+    # -2.7176,
+    # -2.8272,
+    # -3.5522,
+]
