+++ tests/test_lb.py
@@ -1,3 +1,25 @@
+# MIT License
+#
+# Copyright (c) 2020 Archis Joglekar
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
 from vlapy.core import lenard_bernstein
 import numpy as np
 
@@ -112,4 +134,4 @@ def test_lenard_bernstein_velocity_zero():
 
     temp_in = np.sum(f * v) * dv
     temp_out = np.sum(f_out * v) * dv
-    np.testing.assert_almost_equal(temp_out, 0., decimal=1)
+    np.testing.assert_almost_equal(temp_out, 0.0, decimal=1)
