### 2020-04-23 #2133 @dpsanders
Hi @joglekara,

I'm well thanks; I hope you are too.

Sorry for the delay in getting to this.
The documentation looks very nice, thanks for doing that. That will definitely make the package more appealing and usable.

I'll start looking for referees.
