### 2020-04-24 #2133 @StanczakDominik
> 👋 Hi @StanczakDominik, @TomGoffrey and @CFGrote
>
> I hope you are all doing OK in these difficult times.
>
> If you have any spare bandwidth, would you be able to review this JOSS submission? Thanks!
>

Wow, sure! It is my first time reviewing (hyped!) and I'd like to make sure I do a good job of it; what are the time frames like? Is it fine if I take a few days to really dig into it?
