### 2020-04-24 #2133 @StanczakDominik
Going through the comments section I see mentions of PlasmaPy. As a on-off maintainer of that package, which I'd describe as "still very smol" - that's a surprise, to be sure, but a welcome one. I don't know if that makes for any sort of conflict of interest, though? I'd be happy to do my best to review this package on a completely stand alone basis.
