### 2020-04-24 #2133 @joglekara
@dpsanders : No problem about the delay at all. The times have not been kind to most this year so thanks a bunch for taking time to get this started.

@StanczakDominik : Appreciate your enthusiasm! @dpsanders is right in that I'm interested in further connecting with PlasmaPy. I think it'd be great to be able to put that package on firm ground.

Also, @TomGoffrey, @StanczakDominik , nice to e-meet you 👋
