### 2020-04-24 898076a
refactoring storage to 1. include some of the low level code 2. support complex valued quantities 3. store fourier modes rather than the whole distribution
also some comments
