+++ vlapy/storage.py
@@ -28,7 +28,24 @@ import numpy as np
 
 
 class StorageManager:
-    def __init__(self, xax, vax, tax, base_path):
+    def __init__(self, xax, vax, tax, base_path, store_f=None):
+        """
+        This is the initialization for the storage class.
+        The paths are set
+
+        At the moment, we store the electric field at every timestep
+        we store the driver field at every timestep
+        and we have the choice of storing the distribution function at every time-step
+        but since it's typically too large, we can store a few choice Fourier modes
+
+        :param xax:
+        :param vax:
+        :param tax:
+        :param base_path:
+        :param store_f:
+        """
+        self.initialize_temporary_storage(xax, vax, tax, store_f)
+
         self.base_path = base_path
         self.efield_path = os.path.join(base_path, "electric_field_vs_time.nc")
         self.driver_efield_path = os.path.join(
@@ -37,12 +54,60 @@ class StorageManager:
         self.f_path = os.path.join(base_path, "dist_func_vs_time.nc")
 
         self.__init_electric_field_storage(tax=tax, xax=xax)
-        self.__init_dist_func_storage(tax=tax, xax=xax, vax=vax)
+
+        self.store_f = store_f
+
+        if store_f is not None:
+            self.__init_dist_func_storage(
+                tax=tax, xax=xax, vax=vax, f_storage_rules=store_f
+            )
 
     def close(self):
-        self.efield_arr.to_netcdf(self.efield_path)
-        self.driver_efield_arr.to_netcdf(self.driver_efield_path)
-        self.f_arr.to_netcdf(self.f_path)
+        """
+        This method closes all the files safely before program termination
+
+        :return:
+        """
+        self.efield_arr.to_netcdf(
+            self.efield_path, engine="h5netcdf", invalid_netcdf=True
+        )
+        self.driver_efield_arr.to_netcdf(
+            self.driver_efield_path, engine="h5netcdf", invalid_netcdf=True
+        )
+        if self.store_f:
+            self.f_arr.to_netcdf(self.f_path, engine="h5netcdf", invalid_netcdf=True)
+
+    def initialize_temporary_storage(self, xax, vax, tax, store_f):
+        """
+        This method initializes storage
+
+        :param xax:
+        :param vax:
+        :param tax:
+        :param store_f:
+        :return:
+        """
+        nt = tax.size
+        nx = xax.size
+        nv = vax.size
+
+        if nt // 4 < 100:
+            self.t_store = 100
+        else:
+            self.t_store = nt // 4
+
+        self.temp_field_store = np.zeros([self.t_store, nx])
+        self.temp_driver_store = np.zeros([self.t_store, nx])
+        self.temp_t_store = np.zeros(self.t_store)
+        self.it_store = 0
+
+        if store_f == "all-x":
+            self.temp_dist_store = np.zeros([self.t_store, nx, nv])
+        else:
+            kax = np.linspace(0, 1, 2)
+            self.temp_dist_store = np.zeros(
+                (self.t_store, kax.size, nv), dtype=np.complex
+            )
 
     def __init_electric_field_storage(self, tax, xax):
         """
@@ -60,9 +125,9 @@ class StorageManager:
             data=electric_field_store, coords=[("time", tax), ("space", xax)]
         )
 
-        ef_DA.to_netcdf(self.efield_path)
+        ef_DA.to_netcdf(self.efield_path, engine="h5netcdf", invalid_netcdf=True)
 
-        self.efield_arr = xr.open_dataarray(self.efield_path)
+        self.efield_arr = xr.load_dataarray(self.efield_path, engine="h5netcdf")
 
[…]
