### 2020-04-24 a5ec659
almost there but `xarray` can't find `h5netcdf` for some reason
