+++ diagnostics/landau_damping.py
@@ -20,72 +20,45 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
+
 import os
 
 import numpy as np
 import mlflow
 from matplotlib import pyplot as plt
 
+from diagnostics import low_level_helpers as llh
+
 
 class LandauDamping:
-    def __init__(self):
-        self.params_to_log = [
-            "k0",
-            "w0",
-            "a0",
-        ]
+    def __init__(self, params_to_log):
+        self.params_to_log = params_to_log
+
+        self.f_rules = "k0k1"
 
     def __call__(self, storage_manager):
         self.plots_dir = os.path.join(storage_manager.base_path, "plots")
         os.makedirs(self.plots_dir)
 
         metrics = {
-            "damping_rate": self.get_damping_rate(storage_manager),
-            "frequency": self.get_oscillation_frequency(storage_manager),
+            "damping_rate": llh.get_damping_rate(efield_arr=storage_manager.efield_arr),
+            "frequency": llh.get_oscillation_frequency(
+                efield_arr=storage_manager.efield_arr
+            ),
+            "E_max": llh.get_e_max(efield_arr=storage_manager.efield_arr),
         }
 
         mlflow.log_metrics(metrics=metrics)
 
         self.make_plots(storage_manager)
 
-    def get_damping_rate(self, storage_manager):
-        efield_arr = storage_manager.efield_arr
-        tax = efield_arr.coords["time"].data
-
-        t_ind = tax.size // 4
-
-        ek = np.fft.fft(efield_arr.data, axis=1)
-        ek_mag = np.array([np.abs(ek[it, 1]) for it in range(tax.size)])[t_ind:]
-
-        dEdt = np.gradient(np.log(ek_mag), tax[2] - tax[1])
-
-        return np.mean(dEdt)
-
-    def get_oscillation_frequency(self, storage_manager):
-        efield_arr = storage_manager.efield_arr
-        tax = efield_arr.coords["time"].data
-        nt = tax.size
-        dt = tax[1] - tax[0]
-
-        ekw = np.fft.fft2(efield_arr.data[nt // 4 :,])
-        ek1w = np.abs(ekw[:, 1])
-
-        wax = np.fft.fftfreq(ekw.shape[0], d=dt) * 2 * np.pi
-
-        return wax[ek1w.argmax()]
-
     def make_plots(self, storage_manager):
-        efield_arr = storage_manager.efield_arr
-        tax = efield_arr.coords["time"].data
-        dt = tax[1] - tax[0]
-        wax = np.fft.fftfreq(tax.size, d=dt) * 2 * np.pi
+        tax = storage_manager.efield_arr.coords["time"].data
+        ek_rec = llh.get_first_mode(storage_manager.efield_arr)
 
-        ek = np.fft.fft(efield_arr.data, axis=1)
-        ek_mag = np.array([np.abs(ek[it, 1]) for it in range(tax.size)])
-        ekw_mag = np.abs(np.fft.fft(np.array([ek[it, 1] for it in range(tax.size)])))
+        ek_mag = np.array([np.abs(ek_rec[it, 1]) for it in range(tax.size)])
 
         self.__plot_e_vs_t(tax, ek_mag, "Electric Field Amplitude vs Time")
-        self.__plot_e_vs_w(wax, ekw_mag, "Electric Field Amplitude vs Frequency")
 
     def __plot_e_vs_t(self, t, e, title):
         this_fig = plt.figure(figsize=(8, 4))
@@ -98,20 +71,7 @@ class LandauDamping:
[…]
