### 2020-04-24 f3b1a8c
h5netcdf errors were local path issues. fixed. CircleCI didn't have those problems
Some of the tests needed rewriting to address the new fourier mode storage feature
