+++ tests/test_storage.py
@@ -37,7 +37,9 @@ def test_storage_init_files_exist():
     dirname = os.path.join(os.getcwd(), str(uuid.uuid4()))
     os.makedirs(dirname, exist_ok=True)
 
-    st = storage.StorageManager(xax=xax, vax=vax, tax=tax, base_path=dirname)
+    st = storage.StorageManager(
+        xax=xax, vax=vax, tax=tax, base_path=dirname, store_f="all-x"
+    )
 
     assert os.path.exists(os.path.join(dirname, "electric_field_vs_time.nc"))
     assert os.path.exists(os.path.join(dirname, "dist_func_vs_time.nc"))
@@ -53,12 +55,36 @@ def test_storage_init_shape():
     dirname = os.path.join(os.getcwd(), str(uuid.uuid4()))
     os.makedirs(dirname, exist_ok=True)
 
-    st = storage.StorageManager(xax=xax, vax=vax, tax=tax, base_path=dirname)
+    st = storage.StorageManager(
+        xax=xax, vax=vax, tax=tax, base_path=dirname, store_f="all-x"
+    )
 
     np.testing.assert_equal(st.efield_arr.coords["space"].size, xax.size)
     np.testing.assert_equal(st.efield_arr.coords["time"].size, tax.size)
 
     np.testing.assert_equal(st.f_arr.coords["space"].size, xax.size)
+    np.testing.assert_equal(st.f_arr.coords["time"].size, tax.size)
+    np.testing.assert_equal(st.f_arr.coords["velocity"].size, vax.size)
+
+    shutil.rmtree(dirname)
+
+
+def test_storage_init_shape_fourier():
+    xax = np.linspace(0, 1, 16)
+    vax = np.linspace(0, 1, 24)
+    tax = np.linspace(0, 1, 32)
+
+    dirname = os.path.join(os.getcwd(), str(uuid.uuid4()))
+    os.makedirs(dirname, exist_ok=True)
+
+    st = storage.StorageManager(
+        xax=xax, vax=vax, tax=tax, base_path=dirname, store_f="k0k1"
+    )
+
+    np.testing.assert_equal(st.efield_arr.coords["space"].size, xax.size)
+    np.testing.assert_equal(st.efield_arr.coords["time"].size, tax.size)
+
+    np.testing.assert_equal(st.f_arr.coords["fourier_mode"].size, 2)
     np.testing.assert_equal(st.f_arr.coords["time"].size, tax.size)
     np.testing.assert_equal(st.f_arr.coords["velocity"].size, vax.size)
 
