### 2020-04-26 #2133 @StanczakDominik
Thanks, both of you! #EqualityForBots
