### 2020-04-26 joglekara/VlaPy#21 @StanczakDominik
Updates for the `Landau_damping` notebook
The notebook needs a few updates:

- [ ] The "add `..` to sys.path"  trick at the start made my import fail. It would also fail with the notebook in any other location. I think if the setup.py install or pip install of the package goes through well, it should be unnecessary.
- [ ] The `from vlapy` import needs to be replaced with `from vlapy.core`.
   - [ ] To prevent this kind of issue, I think the notebook could be put into CI and rendered via nbsphinx through readthedocs; [I recently did that for PlasmaPy](https://github.com/PlasmaPy/PlasmaPy/pull/792) and could contribute that. There's also a way to build the docs for every PR; ~~it was somewhere along here https://github.com/orgs/readthedocs/projects/8 . I'll try to recall where the sign up for that is.~~ According to https://docs.readthedocs.io/en/stable/guides/autobuild-docs-for-pull-requests.html you can sign up for it via email.
- [ ] The notebook - especially the "Initialize and run" part - could definitely use some more breaking up to explain the defined variables, especially the commented out sets of those. Drop some more narrative in!

* A question, because the docs weren't really clear on that - what's the connection between the notebook and the MLFlow style run? does the notebook implement the same sort of simulation?
* A stretch goal - I would have loved an animation to go along with the 2D plot ([`xarray.interactive`](https://github.com/pydata/xarray/issues/3709) may soon be an option to do that). That's completely optional, though.
