### 2020-04-26 joglekara/VlaPy#21 @joglekara
Thanks for the feedback here.

This notebook is quite a relic at this point. I made it early on in an effort to clarify the API but it's beyond needing a simple update or two, as you rightfully have pointed out.

I vote to remove it. What do you think?
