+++ .readthedocs.yaml
@@ -21,5 +21,5 @@ python:
   version: 3.7
   install:
     - method: setuptools
-      path: vlapy
+      path: .
   system_packages: true
