+++ .readthedocs.yaml
@@ -20,6 +20,11 @@ formats: all
 python:
   version: 3.7
   install:
+    - requirements: docs/requirements.txt
+    - method: pip
+      path: .
+      extra_requirements:
+        - docs
     - method: setuptools
       path: .
   system_packages: true
