### 2020-04-26 872254e
removed tracking uri option because it wasn't working for local locations anyway (mlflow bug)
and added screen output to provide some feedback
