+++ vlapy/manager.py
@@ -24,6 +24,7 @@ import os
 import uuid
 
 import shutil
+from tqdm import tqdm
 import mlflow
 import numpy as np
 
@@ -189,7 +190,7 @@ def start_run(all_params, pulse_dictionary, diagnostics, name="test"):
         )
 
         # Time Loop
-        for it in range(nt):
+        for it in tqdm(range(nt)):
             e, f = step.full_PEFRL_ps_step(
                 f, x, kx, v, kv, dv, t[it], dt, e, driver_function
             )
