### 2020-04-26 joglekara/VlaPy#21 @joglekara
Here's a slightly more updated notebook but still needs to be brought up to full speed:
https://github.com/joglekara/VlaPy/tree/ld_notebook_mlflow/notebooks

Would be happy to have a contributor on it, though!
