### 2020-04-27 joglekara/VlaPy#21 @StanczakDominik
The other notebook you've linked is really lovely in terms of displaying the current API! I think taking the explanations from the first one and adding some explanations for each of the physics parameters makes this a great downloadable introduction.

I'll try to find something in the next few days to procrastinate away by adding nbsphinx to the docs :laughing:
