### 2020-04-27 760e7c2
Apply suggestions from code review
Adding suggested pretty print

Co-Authored-By: Dominik Stańczak <stanczakdominik@gmail.com>
