+++ run_vlapy.py
@@ -22,7 +22,7 @@
 
 import os
 from datetime import datetime
-
+import pprint
 
 import numpy as np
 
@@ -67,9 +67,9 @@ if __name__ == "__main__":
     print("MLFlow experiment name: " + mlflow_exp_name)
     print("mlruns folder located at " + os.getcwd())
     print("Run parameters: ")
-    print(all_params_dict)
+    pprint.pprint(all_params_dict)
     print("Driver parameters: ")
-    print(pulse_dictionary)
+    pprint.pprint(pulse_dictionary)
     print()
     print(
         "run `mlflow ui` at the command line, and go "
