### 2020-04-27 c0efe44
added docs for parameters for core functions
