### 2020-04-27 577e7b8
added docs for parameters for rest of functions in main folder
