+++ vlapy/manager.py
@@ -20,10 +20,8 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-import os
-import uuid
+import tempfile
 
-import shutil
 from tqdm import tqdm
 import mlflow
 import numpy as np
@@ -118,97 +116,92 @@ def start_run(all_params, pulse_dictionary, diagnostics, name="test"):
     mlflow.set_experiment(name)
 
     with mlflow.start_run():
-        # Log desired parameters
-        params_to_log_dict = {}
-        for param in diagnostics.params_to_log:
-            if param in ["a0", "k0", "w0"]:
-                params_to_log_dict[param] = pulse_dictionary["first pulse"][param]
-            else:
-                params_to_log_dict[param] = all_params[param]
+        with tempfile.TemporaryDirectory() as temp_path:
+            # Log desired parameters
+            params_to_log_dict = {}
+            for param in diagnostics.params_to_log:
+                if param in ["a0", "k0", "w0"]:
+                    params_to_log_dict[param] = pulse_dictionary["first pulse"][param]
+                else:
+                    params_to_log_dict[param] = all_params[param]
 
-        mlflow.log_params(params_to_log_dict)
+            mlflow.log_params(params_to_log_dict)
 
-        # Initialize machinery
-        nx = all_params["nx"]
-        nv = all_params["nv"]
-        nu = all_params["nu"]
-        tmax = all_params["tmax"]
-        nt = all_params["nt"]
+            # Initialize machinery
+            nx = all_params["nx"]
+            nv = all_params["nv"]
+            nu = all_params["nu"]
+            tmax = all_params["tmax"]
+            nt = all_params["nt"]
 
-        # Distribution function
-        f = step.initialize(nx, nv)
+            # Distribution function
+            f = step.initialize(nx, nv)
 
-        # Spatial Grid
-        # Fixed to single wavenumber domains
-        xmax = all_params["xmax"]
-        xmin = all_params["xmin"]
-        dx = (xmax - xmin) / nx
-        x = np.linspace(xmin + dx / 2.0, xmax - dx / 2.0, nx)
-        kx = np.fft.fftfreq(x.size, d=dx) * 2.0 * np.pi
+            # Spatial Grid
+            xmax = all_params["xmax"]
+            xmin = all_params["xmin"]
+            dx = (xmax - xmin) / nx
+            x = np.linspace(xmin + dx / 2.0, xmax - dx / 2.0, nx)
+            kx = np.fft.fftfreq(x.size, d=dx) * 2.0 * np.pi
 
-        # Velocity grid
-        vmax = all_params["vmax"]
-        dv = 2 * vmax / nv
-        v = np.linspace(-vmax + dv / 2.0, vmax - dv / 2.0, nv)
-        kv = np.fft.fftfreq(v.size, d=dv) * 2.0 * np.pi
+            # Velocity grid
+            vmax = all_params["vmax"]
+            dv = 2 * vmax / nv
+            v = np.linspace(-vmax + dv / 2.0, vmax - dv / 2.0, nv)
+            kv = np.fft.fftfreq(v.size, d=dv) * 2.0 * np.pi
 
-        t = np.linspace(0, tmax, nt)
-        dt = t[1] - t[0]
+            t = np.linspace(0, tmax, nt)
+            dt = t[1] - t[0]
 
-        def driver_function(x, tt):
-            total_field = np.zeros_like(x)
+            def driver_function(x, tt):
+                total_field = np.zeros_like(x)
 
-            for this_pulse in list(pulse_dictionary.keys()):
[…]
