### 2020-04-28 c3c473d
couple of bugs in docs
