### 2020-05-11 631e8fb
new diagnostics
much slower, (~15x)
has some ringing
