+++ diagnostics/landau_damping.py
@@ -34,7 +34,7 @@ class LandauDamping:
     def __init__(self, params_to_log):
         self.params_to_log = params_to_log
 
-        self.f_rules = "k0k1"
+        self.f_rules = "all-x"
 
     def __call__(self, storage_manager):
         self.plots_dir = os.path.join(storage_manager.base_path, "plots")
