### 2020-05-17 joglekara/VlaPy#31 @StanczakDominik
Better describe the goal, target audience of the project
Hey, I'm getting back to reviewing the project (sorry for the delay!). I got to this point:

>  A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

And in the documentation and README I can see only

> VlaPy is a 1-spatial-dimension, 1-velocity-dimension, Vlasov-Poisson-Fokker-Planck code written in Python. The Vlasov-Poisson-Fokker-Planck system of equations is commonly used in plasma physics.

I think this could be expanded a little. The goal is stated pretty clearly, but the target audience is not mentioned, except very broadly (plasma physicists).
