### 2020-05-17 joglekara/VlaPy#32 @StanczakDominik
Contributor instructions
Going through the JOSS review, looking at https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines

> There should be clear guidelines for third-parties wishing to:
>   Contribute to the software
>   Report issues or problems with the software
>   Seek support

And I'm not really seeing anything like that here. It's probably a good idea to include. Perhaps something like https://github.com/sunpy/sunpy/blob/master/CONTRIBUTING.rst could be a good idea?
