### 2020-05-17 joglekara/VlaPy#32 @joglekara
Thanks for the feedback! Will implement
