### 2020-05-17 joglekara/VlaPy#31 @joglekara
Thanks for the feedback. Will expand on that a bit more.
