### 2020-05-17 joglekara/VlaPy#31 @StanczakDominik
I just got to the actual paper, and the section in there sounds good :)
