### 2020-05-17 #2182 @StanczakDominik
For the "State of the field: Do the authors describe how this software compares to other commonly-used packages?" checklist item, I think this section should be sufficient:

> There are many software libraries that solve the same equation set which are available in academic settings, research laboratories, and industry (e.g., [@Banks2017; @Joglekar2018]), but the community has yet to benefit from a simple-to-read, open-source Python implementation. This lack of capability is currently echoed in conversations within the PlasmaPy [@plasmapy] community (PlasmaPy is a collection of open-source plasma physics resources). Our aim with VlaPy is to take a step towards filling this need in the open-source community.
