### 2020-05-17 joglekara/VlaPy#21 @StanczakDominik
Plasmapy's PR 792 is still stalled on review, but I think I could quickly add nbsphinx here, if you're down for that :smiley:
