### 2020-05-17 #2182 @StanczakDominik
- "A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?" is being dealt with by https://github.com/joglekara/VlaPy/issues/31
- "Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support", by https://github.com/joglekara/VlaPy/issues/32
- "Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).", by https://github.com/joglekara/VlaPy/issues/21
