### 2020-05-19 joglekara/VlaPy#31 @joglekara
Ah, thanks!

In that case, should we close this?

Or do you believe I should add that to the README as well? It doesn't seem like most repositories do that but I'm happy to be proven wrong!
