### 2020-05-19 joglekara/VlaPy#21 @joglekara
@StanczakDominik let's do it!
