### 2020-05-20 joglekara/VlaPy#31 @StanczakDominik
I think it'd make sense for it to be in the readme as well. That's the first place someone stumbling upon a project looks at.
