### 2020-05-20 e6a626a
brought in text from manuscript to statement of need in readme
