### 2020-05-25 a95e9d0
updated the landau damping notebook to clarify API
