### 2020-05-25 joglekara/VlaPy#21 @StanczakDominik
I merged nbsphinx into plasmapy yesterday; now that I got a second opinion there that I'm not doing anything wildly unreasonable, I feel more comfortable adding it here. I'll get it done by tomorrow - likely today.
