### 2020-05-29 f5f8ff0
added references for research applications in LPI, Accelerator, Space, and applied math/fundamental plasma physics.
