### 2020-05-29 joglekara/VlaPy#21 @StanczakDominik
Hey, re: nbsphinx, I've been thinking about it a bit more and realized that the execute-notebook-at-doc-build part is not going to play nicely with MLFlow having to be launched before execution - its only use would be to include the already built notebook, but there's strong risk of the notebook diverging from the main API.
