### 2020-05-29 joglekara/VlaPy#21 @joglekara
ah, np. Regardless, thanks for looking into it @StanczakDominik !
