### 2020-05-30 #2182 @joglekara
> My apologies in the delay on this review.
>
> I'm currently stalling at the statement of need and example usage requirements.
>
> Reading the paper it's not clear to me what the primary aim of the software is. Is it intended as a research tool in its own right, an educational tool, or both? I think this could be clarified, and assuming not a purely educational tool some concrete references to example applications using 1D1V Vlasov would ideally be added.
>
> I have a couple of other minor comments at this stage:
>
> * Is the exclusion of sphinx, sphinxcontrib-bibtex and sphinx_rtd_theme from the INSTALL_REQUIREMENTS in setup.py deliberate?
> * Personally I would add newline at the end of README.md and CONTRIBUTING.md, as is done in CODE_OF_CONDUCT.md

The statement of need/applications is hopefully addressed in https://github.com/joglekara/VlaPy/pull/38. Looking forward to your feedback there @TomGoffrey . The newline suggestions are also included there.

I suppose the main reason sphinx and the related packages are excluded is because we rely on the `.read_the_docs.yaml` and `docs/requirements.txt` to do the work. However, that's only configured for hosting on RTD. Do you think it be added to the `setup.py` for the main package too?
