### 2020-06-29 #2182 @dpsanders
Thanks @TomGoffrey! My opinion is that simple modifications to the paper and tests that would make the paper and repo more accessible should definitely be done.
