### 2020-06-30 #2182 @joglekara
Thanks @TomGoffrey for the suggestions!

> Yes, the statement of need is much clearer now thanks.
>
> Regarding the sphinx requirements discussed above, I'm afraid I don't have a definitive opinion. On one hand my guess would be that not many users would build the docs locally, but on the other hand it's not a big change and it would benefit the users who do want to build for the requirements to be added. That all being said as I don't have a strong opinion I'm happy to leave the decision to you.
>
> Some additional comments:
>
> 1. I think the paper itself would benefit from a plot (or two) for the final section. I think you could easily adapt the later plots from the notebook? Obviously with the notebook the user could get the plots that way, but I think my adding it to the paper you make it a stronger stand-alone document.
Agreed! I will implement that.

> 2. The other tests are mentioned, but it's not obvious how to run them. Obviously it doesn't take much searching through the test directory, but a simple comment along the lines of these tests are demonstrated in `tests/test_lb.py` would be beneficial. Even better (but I guess more work) would be (e.g.) a plot showing the relaxation to a Maxwellian. Again, sounds basic but in my experience many first time users like to be able to verify they have run the code correctly, and regenerating published plots is an obvious starting point.
Good idea. Implementing this in a Jupyter Notebook seems like the best way forward. Do you agree?

> 3. Is there a reason many of the RTD pages end in '…this page is under development…'? Is there more to do?
Well, there's always more to do ;). For many of those notes, I was referring to the fact that the the background to some of the finite differencing or kinetic equation derivations is still a work in progress.

Perhaps it is better to clean up any loose ends, trim and close the loop? Or leave the loop open while explicitly stating what is still to be done? @TomGoffrey @StanczakDominik @dpsanders , your thoughts are welcome here.

>
> It's possible some (all?) of the above is considered beyond the required scope of the article, as I think I could reasonably sign off on the checklist now. @dpsanders could you comment perhaps?
