### 2020-07-01 c113026
added a screenshot and edits to the RTD
