### 2020-07-03 71a1cd7
added daugherty method
have to fix a bunch of tests
