+++ tests/test_dg.py
@@ -24,7 +24,7 @@ from vlapy.core import collisions
 import numpy as np
 
 
-def test_lenard_bernstein_maxwellian_solution():
+def test_dg_maxwellian_solution():
     """
     tests if df/dt = 0 if f = maxwellian
 
@@ -37,20 +37,19 @@ def test_lenard_bernstein_maxwellian_solution():
 
     nu = 1e-3
     dt = 0.1
-    v0 = 1.0
 
-    leftside = collisions.make_philharmonic_matrix(v, nv, nu, dt, dv, v0)
-
-    f = np.exp(-(v ** 2.0) / 2.0 / v0) / np.sum(np.exp(-(v ** 2.0) / 2.0 / v0) * dv)
+    f = np.exp(-(v ** 2.0))
+    f = f / np.sum(f * dv)
     f_out = f.copy()
 
     for it in range(32):
+        leftside = collisions.make_daugherty_matrix(v, nv, nu, dt, dv, f_v=f_out)
         f_out = collisions.take_collision_step(leftside, f_out)
 
-    np.testing.assert_almost_equal(f, f_out, decimal=4)
+    np.testing.assert_almost_equal(f, f_out, decimal=5)
 
 
-def test_lenard_bernstein_energy_conservation():
+def test_dg_energy_conservation():
     """
     tests if the 2nd moment of f is conserved
 
@@ -63,23 +62,21 @@ def test_lenard_bernstein_energy_conservation():
 
     nu = 1e-3
     dt = 0.01
-    v0 = 1.0
 
-    leftside = collisions.make_philharmonic_matrix(v, nv, nu, dt, dv, v0)
-
-    f = np.exp(-((v - 0.5) ** 2.0) / 2.0 / v0)
+    f = np.exp(-((v - 0.5) ** 2.0))
     f = f / np.sum(f * dv)
-
     f_out = f.copy()
+
     for it in range(32):
+        leftside = collisions.make_daugherty_matrix(v, nv, nu, dt, dv, f_v=f_out)
         f_out = collisions.take_collision_step(leftside, f_out)
 
     temp_in = np.sum(f * v ** 2.0) * dv
     temp_out = np.sum(f_out * v ** 2.0) * dv
-    np.testing.assert_almost_equal(temp_out, temp_in, decimal=3)
+    np.testing.assert_almost_equal(temp_out, temp_in, decimal=6)
 
 
-def test_lenard_bernstein_density_conservation():
+def test_dg_density_conservation():
     """
     tests if the 0th moment of f is conserved
 
@@ -92,23 +89,21 @@ def test_lenard_bernstein_density_conservation():
 
     nu = 1e-3
     dt = 0.1
-    v0 = 1.0
 
-    leftside = collisions.make_philharmonic_matrix(v, nv, nu, dt, dv, v0)
-
-    f = np.exp(-((v - 0.5) ** 2.0) / 2.0 / v0)
+    f = np.exp(-((v - 0.5) ** 2.0))
     f = f / np.sum(f * dv)
-
     f_out = f.copy()
+
     for it in range(32):
+        leftside = collisions.make_daugherty_matrix(v, nv, nu, dt, dv, f_v=f_out)
         f_out = collisions.take_collision_step(leftside, f_out)
 
     temp_in = np.sum(f) * dv
     temp_out = np.sum(f_out) * dv
-    np.testing.assert_almost_equal(temp_out, temp_in, decimal=3)
+    np.testing.assert_almost_equal(temp_out, temp_in, decimal=6)
 
 
-def test_lenard_bernstein_velocity_zero():
+def test_dg_velocity_zero():
     """
[…]
