+++ .circleci/config.yml
@@ -8,6 +8,7 @@ jobs:
     # Run the steps with Docker
     docker:
       - image: cimg/python:3.7
+    parallelism: 2
 
       # Specify service dependencies here if necessary
       # CircleCI maintains a library of pre-built images
