### 2020-07-03 4708a10
updated collision notebook
