+++ tests/test_dg.py
@@ -37,17 +37,18 @@ def test_dg_maxwellian_solution():
 
     nu = 1e-3
     dt = 0.1
+    nt = 8
 
     f = np.exp(-(v ** 2.0))
     f = f / np.sum(f * dv)
     f_out = f.copy()
 
-    for it in range(8):
+    for it in range(nt):
         f_out = collisions.take_collision_step(
             collisions.make_daugherty_matrix, f_out, v, nv, nu, dt, dv
         )
 
-    np.testing.assert_almost_equal(f, f_out, decimal=5)
+    np.testing.assert_almost_equal(f, f_out, decimal=3)
 
 
 def test_dg_energy_conservation():
@@ -63,12 +64,13 @@ def test_dg_energy_conservation():
 
     nu = 1e-3
     dt = 0.01
+    nt = 8
 
     f = np.exp(-((v - 0.5) ** 2.0))
     f = f / np.sum(f * dv)
     f_out = f.copy()
 
-    for it in range(8):
+    for it in range(nt):
         f_out = collisions.take_collision_step(
             collisions.make_daugherty_matrix, f_out, v, nv, nu, dt, dv
         )
@@ -91,12 +93,13 @@ def test_dg_density_conservation():
 
     nu = 1e-3
     dt = 0.1
+    nt = 8
 
     f = np.exp(-((v - 0.5) ** 2.0))
     f = f / np.sum(f * dv)
     f_out = f.copy()
 
-    for it in range(8):
+    for it in range(nt):
         f_out = collisions.take_collision_step(
             collisions.make_daugherty_matrix, f_out, v, nv, nu, dt, dv
         )
@@ -106,29 +109,30 @@ def test_dg_density_conservation():
     np.testing.assert_almost_equal(temp_out, temp_in, decimal=6)
 
 
-# def test_dg_velocity_zero():
-#     """
-#     tests if the 1st moment of f is (approximately) 0
-#
-#     :return:
-#     """
-#     vmax = 6.0
-#     nv = 256
-#     dv = 2 * vmax / nv
-#     v = np.linspace(-vmax + dv / 2.0, vmax - dv / 2.0, nv)
-#
-#     nu = 5e-2
-#     dt = 0.1
-#
-#     f = np.exp(-((v - 0.1) ** 2.0))
-#     f = f / np.sum(f * dv)
-#     f_out = f.copy()
-#
-#     for it in range(50000):
-#         f_out = collisions.take_collision_step(
-#             collisions.make_daugherty_matrix, f_out, v, nv, nu, dt, dv
-#         )
-#
-#     temp_in = np.sum(f * v) * dv
-#     temp_out = np.sum(f_out * v) * dv
-#     np.testing.assert_almost_equal(temp_out, 0.0, decimal=4)
+def test_dg_momentum_conservation():
+    """
+    tests if the 0th moment of f is conserved
+
+    :return:
+    """
+    vmax = 6.0
+    nv = 512
+    dv = 2 * vmax / nv
+    v = np.linspace(-vmax + dv / 2.0, vmax - dv / 2.0, nv)
+    nt = 8
+
+    nu = 1e-3
+    dt = 0.1
+
+    f = np.exp(-((v - 0.5) ** 2.0))
+    f = f / np.sum(f * dv)
+    f_out = f.copy()
+
+    for it in range(nt):
+        f_out = collisions.take_collision_step(
+            collisions.make_daugherty_matrix, f_out, v, nv, nu, dt, dv
+        )
+
+    temp_in = np.sum(f * v) * dv
+    temp_out = np.sum(f_out * v) * dv
+    np.testing.assert_almost_equal(temp_out, temp_in, decimal=6)
