### 2020-07-03 #2182 @TomGoffrey
@joglekara I do agree that a notebook would be the best way, looks like you're already working on it. That's great.

I agree, documentation is always WIP, but (and I guess this makes it more your decision than mine) I think the code looks better to (prospective) users if you tie up any loose ends and remove the comments. You can always expand on details later. That's just my opinion though, so as you say if others want to weigh in that would be good.
