+++ tests/test_lb.py
@@ -137,30 +137,30 @@ def test_lenard_bernstein_momentum_conservation_if_initialized_at_zero():
     np.testing.assert_almost_equal(temp_out, temp_in, decimal=3)
 
 
-def test_lenard_bernstein_velocity_zero():
-    """
-    tests if the 1st moment of f is (approximately) 0
-
-    :return:
-    """
-    vmax = 6.0
-    nv = 256
-    dv = 2 * vmax / nv
-    v = np.linspace(-vmax + dv / 2.0, vmax - dv / 2.0, nv)
-
-    nu = 5e-2
-    dt = 0.1
-    v0 = 1.0
-
-    f = np.exp(-((v - 0.25) ** 2.0) / 2.0 / v0)
-    f = f / np.sum(f * dv)
-
-    f_out = f.copy()
-    for it in range(500):
-        f_out = collisions.take_collision_step(
-            collisions.make_philharmonic_matrix, f_out, v, nv, nu, dt, dv
-        )
-
-    temp_in = np.sum(f * v) * dv
-    temp_out = np.sum(f_out * v) * dv
-    np.testing.assert_almost_equal(temp_out, 0.0, decimal=1)
+# def test_lenard_bernstein_velocity_zero():
+#     """
+#     tests if the 1st moment of f is (approximately) 0
+#
+#     :return:
+#     """
+#     vmax = 6.0
+#     nv = 256
+#     dv = 2 * vmax / nv
+#     v = np.linspace(-vmax + dv / 2.0, vmax - dv / 2.0, nv)
+#
+#     nu = 5e-2
+#     dt = 0.1
+#     v0 = 1.0
+#
+#     f = np.exp(-((v - 0.25) ** 2.0) / 2.0 / v0)
+#     f = f / np.sum(f * dv)
+#
+#     f_out = f.copy()
+#     for it in range(8):
+#         f_out = collisions.take_collision_step(
+#             collisions.make_philharmonic_matrix, f_out, v, nv, nu, dt, dv
+#         )
+#
+#     temp_in = np.sum(f * v) * dv
+#     temp_out = np.sum(f_out * v) * dv
+#     np.testing.assert_almost_equal(temp_out, 0.0, decimal=1)
