### 2020-07-04 16eac0a
letting test pass because of timing also changed variable names in notebook
