### 2020-07-04 7556ede
figures and DG op in paper
