+++ paper/paper.md
@@ -95,15 +95,26 @@ These are simply unit tests against analytical solutions of integrals of periodi
 Below, we provide an illustration of a manual validation of the Poisson equation solver. These are also provided in 
 `notebooks/test_poisson.ipynb`
 
+![](../notebooks/screenshots_for_example/poisson_solver.png)
+
+
 ## Fokker-Planck Equation
 
-We use a simplified version of the full Fokker-Planck operator [@Lenard1958]. This is given by
+We have implemented two simplified versions of the full Fokker-Planck operator [@Lenard1958, @Daugherty1964]. 
 
+The first of these implementations has the governing equation given by
 $$\left(\frac{\delta f}{\delta t}\right)_{\text{coll}} = \nu \frac{\partial}{\partial v} \left ( v f + v_0^2 \frac{\partial f}{\partial v}\right), $$
-where $v_0$ is the thermal velocity associated with the Maxwell-Boltzmann distribution that is a solution to this equation.
+where $v_0 = \int v^2 f(x,v) dv, $ is the thermal velocity of the distribution. 
+
+The second of these implementations has a governing equation given by
+$$\left(\frac{\delta f}{\delta t}\right)_{\text{coll}} = \nu \frac{\partial}{\partial v} \left ( (v-\underline{v}) f + v_{t}^2 \frac{\partial f}{\partial v}\right), $$
+where $\underline{v} = \int v f(x,v) dv,$ is the mean velocity of the distribution and $v_{t} = \int ((v-\bar{v})^2 f(x,v) dv, $ is the thermal velocity of the shifted distribution.
+
+The second implementation is an extension of the first, and extends momentum conservation for distributions that have a non-zero mean velocity. 
 
 We discretize this backward-in-time, centered-in-space. This procedure results in the time-step scheme given by
-$$ f^{n} = {\Delta t} \nu \left[\left(-\frac{v_0^2}{\Delta v^2} + \frac{1}{2\Delta v}\right) v_{j+1}f^{n+1}_{j+1} + \left(1+2\frac{v_0^2}{\Delta v^2}\right) f^{n+1}_j + \left(-\frac{v_0^2}{\Delta v^2} - \frac{1}{2\Delta v}\right) v_{j-1}f^{n+1}_{j-1}  \right]. $$ 
+$$ f^{n} = \left[{\Delta t} \nu \left(-\frac{v_{0,t}^2}{\Delta v^2} + \frac{1}{2\Delta v}\right) \bar{v}_{j+1}f^{n+1}_{j+1} + \left(1+2{\Delta t} \nu \frac{v_{0,t}^2}{\Delta v^2}\right) f^{n+1}_j + {\Delta t} \nu \left(-\frac{v_{0,t}^2}{\Delta v^2} - \frac{1}{2\Delta v}\right) \bar{v}_{j-1}f^{n+1}_{j-1}  \right]. $$ 
+where $\bar{v} = v$ or $\bar{v} = v - \underline{v}$ depending on the implementation. 
 
 This forms a tridiagonal system of equations that can be directly inverted.
 
@@ -114,13 +125,21 @@ Unit tests are provided for this operator. They can be found in `tests/test_lb.p
 
 2. The operator conserves number density, momentum, and energy when initialized with a zero net velocity.
 
-The `notebooks/test_fokker_planck.ipynb` notebook contains illustrations and examples for these tests. Below, we show 
-how the implementation relaxes a distribution with a small but visible sinusoidal perturbation. This particular 
-example recovers the number density, momentum, and energy to $10^{-6}$ accuracy over 10000 evaluations.
[…]
