+++ paper/paper.bib
@@ -12,6 +12,21 @@ volume = {96},
 year = {2017}
 }
 
+@article{Dougherty1964,
+abstract = {A model Fokker-Planck equation for a single species of particles in a plasma is discussed. This equation has several properties in common with the real equation, and ascribes an approximately correct behavior to most of the particles, though incorrect for the high-energy tail in the thermal distribution. It is shown that the equation can be solved completely, for the small perturbations of a uniform plasma by electric fields harmonic in space and time, in an external magnetic field. Applications to ionospheric radar scattering are briefly discussed; it is shown that in certain circumstances, ion-ion collisions can have a profound effect on the scattering even though the collision frequency is much smaller than the ion gyrofrequency, and this appears to agree with observation.},
+author = {Dougherty, J. P.},
+doi = {10.1063/1.2746779},
+issn = {00319171},
+journal = {Physics of Fluids},
+number = {11},
+pages = {1788},
+title = {{Model Fokker-Planck Equation for a Plasma and Its Solution}},
+url = {https://aip.scitation.org/doi/10.1063/1.2746779},
+volume = {7},
+year = {1964}
+}
+
+
 @article{Joglekar2018,
 author = {Joglekar, Archis S and Winjum, Benjamin J and Tableman, Adam and Wen, Han and Tzoufras, Michail and Mori, Warren B},
 doi = {10.1088/1361-6587/aab978},
