+++ paper/paper.bib
@@ -26,7 +26,6 @@ volume = {7},
 year = {1964}
 }
 
-
 @article{Joglekar2018,
 author = {Joglekar, Archis S and Winjum, Benjamin J and Tableman, Adam and Wen, Han and Tzoufras, Michail and Mori, Warren B},
 doi = {10.1088/1361-6587/aab978},
