### 2020-07-05 cd16518
small edits
