### 2020-07-05 4608e82
updates to usage details
