### 2020-07-05 f061946
update to FP docs w/ figs
