### 2020-07-07 e7cef32
bug in equation but maybe in citation
