+++ paper/paper.md
@@ -100,10 +100,10 @@ Below, we provide an illustration of a manual validation of the Poisson equation
 
 ## Fokker-Planck Equation
 
-We have implemented two simplified versions of the full Fokker-Planck operator [@Lenard1958, @Dougherty1964]. 
+We have implemented two simplified versions of the full Fokker-Planck operator [@Lenard1958; @Dougherty1964]. 
 
 The first of these implementations (LB) has the governing equation given by
-$$\left(\frac{\delta f}{\delta t}\right)_{\text{coll}} = \nu \frac{\partial}{\partial v} \left ( v f + v_0^2 \frac{\partial f}{\partial v}\right), $$
+$$\left(\frac{\delta f}{\delta t}\right)_{\text{coll}} = \nu \frac{\partial}{\partial v} \left( v f + v_0^2 \frac{\partial f}{\partial v}\right), $$
 where $v_0 = \int v^2 f(x,v) dv, $ is the thermal velocity of the distribution. 
 
 The second of these implementations (DG) has a governing equation given by
