+++ tests/test_lb.py
@@ -144,6 +144,8 @@ def test_lenard_bernstein_velocity_zero():
     :return:
     """
     pass
+
+
 #     vmax = 6.0
 #     nv = 256
 #     dv = 2 * vmax / nv
