### 2020-07-07 #2182 @dpsanders
At a guess, try removing the comma and space at the end of  `$v_0 = \int v^2 f(x,v) dv, $`.
The procedure of parsing the LaTeX equations out of the markdown source is delicate.

Also cc-ing @arfon since

`@whedon generate pdf from branch paper`

[didn't seem to work](https://github.com/openjournals/joss-reviews/issues/2182#issuecomment-654553805)
