### 2020-07-07 #2182 @joglekara
i actually think it's unable to render `$\int$` inline, it seems to move on to the next inline `$\int$` call and get stuck there when i remove the first one. I'm pushing a version with all integrals in their `$$` blocks
