### 2020-07-07 #2182 @dpsanders
I suggest using vector graphics (PDF) in the figures rather than PNG, if possible, so that the resolution is higher.
