### 2020-07-07 17b41b8
pdf plots and some other cleanup
