### 2020-07-07 #2182 @joglekara
Had to convert one more plot.
