+++ paper/paper.bib
@@ -167,24 +167,6 @@ volume = {12},
 year = {2016}
 }
 
-
-@article{Chen2019,
-abstract = {How turbulent energy is dissipated in weakly collisional space and astrophysical plasmas is a major open question. Here, we present the application of a field-particle correlation technique to directly measure the transfer of energy between the turbulent electromagnetic field and electrons in the Earth's magnetosheath, the region of solar wind downstream of the Earth's bow shock. The measurement of the secular energy transfer from the parallel electric field as a function of electron velocity shows a signature consistent with Landau damping. This signature is coherent over time, close to the predicted resonant velocity, similar to that seen in kinetic Alfven turbulence simulations, and disappears under phase randomisation. This suggests that electron Landau damping could play a significant role in turbulent plasma heating, and that the technique is a valuable tool for determining the particle energisation processes operating in space and astrophysical plasmas},
-archivePrefix = {arXiv},
-arxivId = {1902.05785},
-author = {Chen, C. H.K. and Klein, K. G. and Howes, G. G.},
-doi = {10.1038/s41467-019-08435-3},
-eprint = {1902.05785},
-issn = {20411723},
-journal = {Nature Communications},
-number = {1},
-publisher = {Springer US},
-title = {{Evidence for electron Landau damping in space plasma turbulence}},
-url = {http://dx.doi.org/10.1038/s41467-019-08435-3},
-volume = {10},
-year = {2019}
-}
-
 @article{Hoyer2017,
 abstract = {xarray is an open source project and Python package that provides a toolkit and data structures for N-dimensional labeled arrays. Our approach combines an application programing interface (API) inspired by pandas with the Common Data Model for self-described scientific data. Key features of the xarray package include label-based indexing and arithmetic, interoperability with the core scientific Python packages (e.g., pandas, NumPy, Matplotlib), out-of-core computation on datasets that don't fit into memory, a wide range of serialization and input/output (I/O) options, and advanced multi-dimensional data manipulation tools such as group-by and resampling. xarray, as a data model and analytics toolkit, has been widely adopted in the geoscience community but is also used more broadly for multi-dimensional data analysis in physics, machine learning and finance.},
 author = {Hoyer, Stephan and Hamman, Joseph J.},
