### 2020-07-07 #2182 @arfon
> Also @arfon, is there an easy way of compiling the PDF locally to iron out these things?

Authors can use the Whedon preview service at http://whedon.theoj.org
