### 2020-07-07 #2182 @TomGoffrey
This looks great, much improved.

A couple more (I think maybe final!) questions on the article:

1. You seem to have two near duplicate paragraphs introducing the equations section. For the sake of completeness you should probably define the terms explicitly, although I realise these are standard definitions.
2. When describing the advection stage, what exactly is the quantity in square brackets you are taking a Fourier transform of? Is the `=` a typo for `+`, or should it be `=-v...`? Should these derivatives be partial derivatives? I think I must be missing something here.
3. I noticed you have a different sign for the Electric field term in your normalised Vlasov equation compared to Cheng & Knorr (see their equation 1a). Have you used a different normalisation, or is this an error? Looking at Cheng & Knorr equation 2b, it could be your form is correct? Additionally comparing your version of Cheng & Knorr equation 1b, you again seem to have a different sign in the electric field. Perhaps this could all be cleared up with a few extra details when you describe the normalised equations.
4. Have you assumed gamma = 1? If so this should be stated.
5. Are you aware of other published results showing the LB operator not conserving momentum for a drifting Maxwellian? If so a reference might put potential users at ease.
