### 2020-07-10 #2182 @joglekara
@TomGoffrey , I've made your proposed changes in 1-4 and provided a response to 5 in the thread for the Pull Request here https://github.com/joglekara/VlaPy/pull/45
