### 2020-07-11 3da1dce
updates to equation etc.
