+++ paper/paper.md
@@ -41,23 +41,25 @@ The details of the ``VlaPy`` implementation are provided in the following sectio
 # Equations
 
 The Vlasov-Poisson-Fokker-Planck system can be decomposed into 4 components. These components, represented using normalized units, are 
-$\tilde{v} = v/v_{th}$, $\tilde{t} = t / \omega_p$, $\tilde{x} = x / (v_{th} / \omega_p)$, $\tilde{m} = m / m_e$, $\tilde{E} = e E / m_e$, $\tilde{f} = f / m n_e v_{th}^3$. The Fourier transform operator is represented by $\mathcal{F}$ and the subscript to the operator indicates the dimension of the transform. 
-
-The Vlasov-Poisson-Fokker-Planck system can be decomposed into 4 components. The normalized quantities are 
-$\tilde{v} = v/v_{th}$, $\tilde{t} = t / \omega_p$, $\tilde{x} = x / (v_{th} / \omega_p)$, $\tilde{m} = m / m_e$, $\tilde{E} = e E / m_e$, $\tilde{f} = f / m n_e v_{th}^3$. The Fourier Transform operator is represented by $\mathcal{F}$. The subscript to the operator indicates the dimension of the transform. 
+$\tilde{v} = v/v_{th}$, $\tilde{t} = t / \omega_p$, $\tilde{x} = x / (v_{th} / \omega_p)$, $\tilde{m} = m / m_e$, $\tilde{E} = e E / m_e$, $\tilde{f} = f / m_e n_e v_{th}^3$. 
+where $v_{th}$ is the thermal velocity, $\omega_p$ is the electron plasma frequency, $m_e$ is the electron mass, and $e$ is the electron charge. 
+The Fourier transform operator is represented by $\mathcal{F}$ and the subscript to the operator indicates the dimension of the transform. 
+$$\mathcal{F}_x (f(x,v)) = \hat{f}(k_x, v) \exp(i k_x x)$$
+$$\mathcal{F}_v (f(x,v)) = \hat{f}(x, k_v) \exp(- i k_v v)$$
+ 
 
 ## Vlasov Equation
 
-The normalized Vlasov equation is given by
-$$ \frac{\partial f}{\partial t} + v  \frac{\partial f}{\partial x} + E \frac{\partial f}{\partial v} = 0 $$.
+The normalized, non-relativistic ($\gamma=1$) Vlasov equation is given by
+$$ \frac{\partial f}{\partial t} + v  \frac{\partial f}{\partial x} - E(x) \frac{\partial f}{\partial v} = 0 $$.
 
 We use operator splitting to advance the time-step [@Cheng1976]. Each one of those operators is then integrated pseudo-spectrally using the following methodology.
 
-We first Fourier transform the operator, as given by 
-$$ \mathcal{F}_x\left[ \frac{d f}{d t} = v \frac{d f}{d x} \right].$$
+We first Fourier transform the advection operator in $\hat{x}$, as given by 
+$$ \mathcal{F}_x\left[ \frac{\partial f}{\partial t} = - v \frac{\partial f}{\partial x} \right].$$
 
-Next, we solve for the change in the plasma distribution function, discretize, and integrate, as given by
-$$\frac{d\hat{f}}{\hat{f}} = v~ (-i k_x)~ dt, $$
+This process enables decoupling of $\hat{x}$ and $\hat{v}$ grids from the time dimension and allows us to write the following as an Ordinary Differential Equation in time for each point in $\hat{v}$. Next, we solve for the change in the plasma distribution function, discretize, and integrate, as given by
[…]
