### 2020-07-12 417149f
updating the discretization and integration documentation
