+++ paper/paper.bib
@@ -26,6 +26,23 @@ volume = {7},
 year = {1964}
 }
 
+@article{Crouseilles2015,
+abstract = {A new splitting is proposed for solving the Vlasov-Maxwell system. This splitting is based on a decomposition of the Hamiltonian of the Vlasov-Maxwell system and allows for the construction of arbitrary high order methods by composition (independent of the specific deterministic method used for the discretization of the phase space). Moreover, we show that for a spectral method in space this scheme satisfies Poisson's equation without explicitly solving it. Finally, we present some examples in the context of the time evolution of an electromagnetic plasma instability which emphasizes the excellent behavior of the new splitting compared to methods from the literature.},
+author = {Crouseilles, Nicolas and Einkemmer, Lukas and Faou, Erwan},
+doi = {10.1016/j.jcp.2014.11.029},
+issn = {00219991},
+journal = {Journal of Computational Physics},
+keywords = {Charge conservation,Hamiltonian splitting,High order time discretization,Vlasov-Maxwell},
+month = {feb},
+number = {279389},
+pages = {224--240},
+publisher = {Elsevier Inc.},
+title = {{Hamiltonian splitting for the Vlasov–Maxwell equations}},
+url = {http://dx.doi.org/10.1016/j.jcp.2014.11.029 https://linkinghub.elsevier.com/retrieve/pii/S0021999114007918},
+volume = {283},
+year = {2015}
+}
+
 @article{Joglekar2018,
 author = {Joglekar, Archis S and Winjum, Benjamin J and Tableman, Adam and Wen, Han and Tzoufras, Michail and Mori, Warren B},
 doi = {10.1088/1361-6587/aab978},
