### 2020-07-14 c9872ce
periods for equations!
