### 2020-07-14 #2182 @joglekara
@TomGoffrey, equation set clarified and a bunch of other edits for clarity and precision. Thanks for this back and forth so far!
