### 2020-07-20 fa32f4c
cite Strang for operator splitting
