### 2020-07-28 750c5ac
signs for collision operator time index superscript
