### 2020-07-31 #2182 @TomGoffrey
Based on that version I only have a couple of comments:

1. t / \omega isn't a dimensionless time. I presume you mean t \omega. I raised this on the PR for the paper branch a couple of weeks ago but perhaps you didn't notice.
2. The Vlasov equation (in the limit of zero B-field) for species is df_\alpha / dt = ... + (q_alpha E / m_alpha) \nabla_\alpha f_alpha. (Sorry I don't now how to render math equations on github.) So if you're using e 'the electron charge' to normalise the electric field, I'm not sure how you end up with a negative E term? Am I missing something, or do you mean e ' the elementary charge'?

If you can check these points I'll give the manuscript a final read through over the weekend, but I think we're there.
