### 2020-08-01 be6031d
wp^{-1} fixing the minus signs
