### 2020-08-04 #2182 @joglekara
> @TomGoffrey thanks for the follow up!
>
> 1. Fixed! Thanks for catching this.  (I've been working off of [joglekara/VlaPy#45](https://github.com/joglekara/VlaPy/pull/45) for the past month, and don't see your comment in there. Perhaps it was in a different PR that got merged, so my apologies.)
> 2. Ah, thanks for that too. You are very right here, and I got lost in making sure the pesky minus signs were consistent between the (tested) code and the (well-known) equations! I finally traced it back to the fact that the FFT/IFFT operation in NumPy is not [symmetric](https://numpy.org/doc/stable/reference/routines.fft.html#module-numpy.fft).

@TomGoffrey , ^^.
