### 2020-08-06 #2182 @TomGoffrey
Sorry, I'm still stuck on the sign issue. You appear to have changed both the sign of the electric field term and the wording of the normalisation. You're now stating (in the description of the normalisation) `and e is the fundamental charge` - what exactly do you mean by `fundamental charge` here? I'd read this as being the same as `elementary charge` i.e. + 1.6e-19 C.

In summary combinations I think are correct:

1. `where e is the charge of an electron ... $ ... + E df / dv$`

or

2. `where e is the elementary charge (1.6e-19 C) ... $ ... - E df / dv`

Does that make sense, or have I confused myself?

Finally, are the axes on the field solver really 'arbitrary'. Naively I'd assume that x, or \tilde{x} would be more appropriate?
