### 2020-08-06 #2182 @joglekara
Hi @TomGoffrey , thanks for this, no reason to apologize! All part of the process in my opinion!

I think that we have to leave the positive sign in `$ ... + E df / dv$` with the `elementary charge = +1.6e-19.` (Changed to elementary because that is more accurate terminology). The reason is because we can't ignore the stationary ion background i.e. `\nabla E = \rho_i + \rho_e`. My understanding is that the minus sign ends up going in front of `rho_e`. Does this make sense and remain self-consistent?

As for the field solver, I left those axes as arbitrary as the test is simply recovering integrals of a periodic function over an arbitrary dimension, rather than specifically working with a spatial coordinate. Do you think it's more illustrative / less distracting if I change those to `x`?
