### 2020-08-06 #2182 @TomGoffrey
@joglekara I'm going to discuss the normalisation, in a belated effort to keep this thread clean.

@dpsanders Just to note, whilst I have finished my checklist I'd like to see this point clarified before signing off on it. Hope that's OK.
