### 2020-08-12 c9d70f0
elementary charge, lambda_D
added a note that this equation is for electrons
