### 2020-08-13 56e3efd
m, q, E, and f normalizations
