### 2020-08-15 0cd9d71
refactoring the manager significantly
offloaded initialization functions to new file

a single run returns the run now for further investigation, this will be used in testing
