+++ vlapy/initializers.py
@@ -1,0 +1,266 @@
+# MIT License
+#
+# Copyright (c) 2020 Archis Joglekar
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
+import mlflow
+import numpy as np
+
+from diagnostics import z_function
+from vlapy.core import step, field, field_driver
+
+
+def initialize_velocity_quantities(vmax, nv):
+    """
+    This function initializes the velocity grid and related quantities
+
+    :param vmax:
+    :param nv:
+    :return:
+    """
+    dv = 2 * vmax / nv
+    v = np.linspace(-vmax + dv / 2.0, vmax - dv / 2.0, nv)
+    kv = np.fft.fftfreq(v.size, d=dv) * 2.0 * np.pi
+
+    return dv, v, kv
+
+
+def initialize_spatial_quantities(xmin, xmax, nx):
+    """
+    This function initializes the spatial grid and related quantities
+
+    :param xmin:
+    :param xmax:
+    :param nx:
+    :return:
+    """
+    dx = (xmax - xmin) / nx
+    x = np.linspace(xmin + dx / 2.0, xmax - dx / 2.0, nx)
+    kx = np.fft.fftfreq(x.size, d=dx) * 2.0 * np.pi
+    one_over_kx = np.zeros_like(kx)
+    one_over_kx[1:] = 1.0 / kx[1:]
+
+    return dx, x, kx, one_over_kx
+
+
+def log_initial_conditions(all_params, pulse_dictionary, diagnostics):
+    """
+    This function logs initial conditions to the mlflow server.
+
+    :param all_params:
+    :param pulse_dictionary:
+    :param diagnostics:
+    :return:
+    """
+    params_to_log_dict = {}
+    for param in diagnostics.params_to_log:
+        if param in ["a0", "k0", "w0"]:
+            params_to_log_dict[param] = pulse_dictionary["first pulse"][param]
+        else:
+            params_to_log_dict[param] = all_params[param]
+
+    mlflow.log_params(params_to_log_dict)
+
+
+def get_everything_ready_for_time_loop(
+    diagnostics, all_params, pulse_dictionary, num_steps
+):
+    """
+    In order to keep the main time loop clean, this function handles all the
+    necessary initialization and array creation for the main simulation time loop.
+
+    Initialized here:
+    spatial grid
+    velocity grid
+    distribution function
+    time grid
+    driver array
+
+    :param diagnostics:
+    :param all_params:
[…]
