### 2020-08-15 fa20980
storage manager refactored
stepper still to go
