+++ vlapy/storage.py
@@ -22,13 +22,75 @@
 
 import os
 import json
+import shutil
 
+import mlflow
 import xarray as xr
 import numpy as np
 
 
+def load_over_all_timesteps(individual_path, overall_path):
+    """
+    This function
+    1 - loads multiple dataarrays
+    2 - concatenates them into one dataset
+    3 - writes to file
+
+    :param individual_path:
+    :param overall_path:
+    :return:
+    """
+    arr = xr.open_mfdataset(
+        individual_path, combine="by_coords", engine="h5netcdf", parallel=True,
+    )
+
+    arr.to_netcdf(
+        overall_path, engine="h5netcdf", invalid_netcdf=True,
+    )
+
+    arr = xr.open_dataset(overall_path, engine="h5netcdf",)
+
+    return arr
+
+
+def get_paths(base_path):
+    """
+    This function writes the paths to a dictionary and also makes the folder structure
+
+    :param base_path:
+    :return:
+    """
+    paths = {
+        "base": base_path,
+        "long_term": os.path.join(base_path, "long_term"),
+        "temp": os.path.join(base_path, "temp"),
+        "e": os.path.join(base_path, "long_term", "electric_field"),
+        "driver": os.path.join(base_path, "long_term", "driver_electric_field"),
+        "distribution": os.path.join(base_path, "long_term", "distribution_function"),
+        "e-individual": os.path.join(base_path, "temp", "electric_field"),
+        "driver-individual": os.path.join(base_path, "temp", "driver_electric_field"),
+        "distribution-individual": os.path.join(
+            base_path, "temp", "distribution_function"
+        ),
+    }
+
+    for key, val in paths.items():
+        os.makedirs(val, exist_ok=True)
+
+    return paths
+
+
 class StorageManager:
-    def __init__(self, x, v, t, base_path, store_f=None):
+    def __init__(
+        self,
+        xax,
+        vax,
+        base_path,
+        num_steps_in_one_loop,
+        all_params,
+        pulse_dictionary,
+        rules_to_store_f=None,
+    ):
         """
         This is the initialization for the storage class.
         The paths are set
@@ -45,23 +107,20 @@ class StorageManager:
         :param v: velocity axis (numpy array of shape (nv,))
         :param t: time axis (numpy array of shape (nt,))
         :param base_path: path to run folder (string)
-        :param store_f: list of Fourier modes to store (len = 0+`num_fourier_modes`)
+        :param rules_to_store_f: list of Fourier modes to store (len = 0+`num_fourier_modes`)
         """
 
-        self.base_path = base_path
-        self.efield_path = os.path.join(base_path, "electric_field_vs_time.nc")
-        self.driver_efield_path = os.path.join(
-            base_path, "driver_electric_field_vs_time.nc"
-        )
-        self.f_path = os.path.join(base_path, "dist_func_vs_time.nc")
+        self.paths = get_paths(base_path)
+        self.rules_to_store_f = rules_to_store_f
+        self.stored = 0
+        self.xax = xax
+        self.vax = vax
+        self.num_timesteps_to_store = num_steps_in_one_loop
 
-        self.initialize_temporary_storage(x, v, t, store_f)
-        self.__init_electric_field_storage(t=t, x=x)
+        self.initialize_storage()
 
-        self.store_f = store_f
-
-        if store_f is not None:
-            self.__init_dist_func_storage(t=t, x=x, v=v, f_storage_rules=store_f)
+        self.write_parameters_to_file(all_params, "all_parameters")
+        self.write_parameters_to_file(pulse_dictionary, "pulses")
 
     def close(self):
         """
@@ -69,181 +128,199 @@ class StorageManager:
 
         :return:
         """
-        self.efield_arr.to_netcdf(
[…]
