+++ diagnostics/base.py
@@ -1,0 +1,142 @@
+# MIT License
+#
+# Copyright (c) 2020 Archis Joglekar
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
+
+import os
+import time
+
+import numpy as np
+from matplotlib import pyplot as plt
+
+
+def __get_figure_and_plot__():
+    this_fig = plt.figure(figsize=(8, 4))
+    this_plt = this_fig.add_subplot(111)
+
+    return this_fig, this_plt
+
+
+def plot_health(health_dir, storage_manager):
+    # t = storage_manager.overall_arrs["e"].coords["time"].data
+    #
+    # for metric, vals in storage_manager.health.items():
+    #     this_fig, this_plt = __get_figure_and_plot__()
+    #     this_plt.plot(t[-vals.size :], vals)
+    #     this_plt.grid()
+    #     this_plt.set_xlabel(r"Time ($\omega_p^{-1}$)", fontsize=12)
+    #     this_plt.set_ylabel(metric, fontsize=12)
+    #     this_plt.set_title(metric + " vs Time", fontsize=14)
+    #     this_fig.savefig(
+    #         os.path.join(health_dir, metric + ".png"), bbox_inches="tight",
+    #     )
+    #     plt.close(this_fig)
+    pass
+
+
+def plot_e_vs_t(plots_dir, t, e, title):
+    this_fig, this_plt = __get_figure_and_plot__()
+    this_plt.plot(t, e)
+    this_plt.set_xlabel(r"Time ($\omega_p^{-1}$)", fontsize=12)
+    this_plt.set_ylabel(r"$\hat{E}_{k=1}$", fontsize=12)
+    this_plt.set_title(title, fontsize=14)
+    this_plt.grid()
+    this_fig.savefig(
+        os.path.join(plots_dir, "E_vs_time.png"), bbox_inches="tight",
+    )
+    plt.close(this_fig)
+
+
+def plot_e_vs_w(plots_dir, w, e, title):
+    this_fig, this_plt = __get_figure_and_plot__()
+    this_plt.semilogy(np.fft.fftshift(w), np.fft.fftshift(e), "-x")
+    this_plt.set_xlabel(r"Frequency ($\omega_p$)", fontsize=12)
+    this_plt.set_ylabel(r"$\hat{\hat{E}}_{k=1}$", fontsize=12)
+    this_plt.set_title(title, fontsize=14)
+    this_plt.grid()
+    this_plt.set_xlim(-5, 5)
[…]
