+++ vlapy/core/collisions.py
@@ -23,59 +23,162 @@
 import numpy as np
 
 
-def make_philharmonic_matrix(v, nv, nu, dt, dv, f_v):
-    """
-    This matrix is composed of the linear operator that must be inverted with respect
-    to the right side, which'll be the distribution function
+def get_philharmonic_matrix_maker(vax, nv, nx, nu, dt, dv):
+    def make_philharmonic_arrays_for_matrix(f_xv):
+        """
+        These are for use with the vectorized collision operator
 
-    :param v: velocity axis (numpy array of shape (nv,))
-    :param nv: size of velocity axis (int)
-    :param nu: collision frequency (float)
-    :param dt: timestep (float)
-    :param dv: velocity-axis spacing (float)
-    :param v0: thermal temperature (float)
-    :return leftside: matrix for the linear operator
-    """
-    v0t_sq = np.trapz(f_v * v ** 2.0, dx=dv, axis=0)
+        :param v: velocity axis (numpy array of shape (nv,))
+        :param nv: size of velocity axis (int)
+        :param nu: collision frequency (float)
+        :param dt: timestep (float)
+        :param dv: velocity-axis spacing (float)
+        :param v0: thermal temperature (float)
+        :return leftside: matrix for the linear operator
+        """
 
-    a = nu * dt * np.ones(nv - 1) * (-v0t_sq / dv ** 2.0 + v[:-1] / 2 / dv)
-    b = nu * dt * np.ones(nv) * (2 * v0t_sq / dv ** 2.0 + 0.0 / 2 / dv)
-    c = nu * dt * np.ones(nv - 1) * (-v0t_sq / dv ** 2.0 - v[1:] / 2 / dv)
-    leftside = np.diag(a, -1) + np.diag(1 + b, 0) + np.diag(c, 1)
+        v0t_sq = np.trapz(f_xv * vax[None,] ** 2.0, dx=dv, axis=1)
 
-    return leftside
+        a = (
+            nu
+            * dt
+            * np.ones((nx, nv - 1))
+            * (-v0t_sq[:, None] / dv ** 2.0 + vax[None, :-1] / 2 / dv)
+        )
+        b = 1.0 + nu * dt * np.ones((nx, nv)) * (2 * v0t_sq[:, None] / dv ** 2.0)
+        c = (
+            nu
+            * dt
+            * np.ones((nx, nv - 1))
+            * (-v0t_sq[:, None] / dv ** 2.0 - vax[None, 1:] / 2 / dv)
+        )
 
+        return a, b, c
 
-def make_dougherty_matrix(v, nv, nu, dt, dv, f_v):
-    """
-    This matrix is composed of the linear operator that must be inverted with respect
-    to the right side, which'll be the distribution function
-
-    :param v: velocity axis (numpy array of shape (nv,))
-    :param nv: size of velocity axis (int)
-    :param nu: collision frequency (float)
-    :param dt: timestep (float)
-    :param dv: velocity-axis spacing (float)
-    :param v0: thermal temperature (float)
-    :return leftside: matrix for the linear operator
-    """
-    vbar = np.trapz(f_v * v, dx=dv, axis=0)
-    v0t_sq = np.trapz(f_v * (v - vbar) ** 2.0, dx=dv, axis=0)
-
-    a = nu * dt * np.ones(nv - 1) * (-v0t_sq / dv ** 2.0 + (v[:-1] - vbar) / 2.0 / dv)
-    b = nu * dt * np.ones(nv) * (2.0 * v0t_sq / dv ** 2.0)
[…]
