### 2020-08-16 8a817e0
collisions refactored both solvers are now tested
