### 2020-08-16 c1aec9c
all tests passing
notebooks next
