+++ vlapy/core/vlasov_poisson.py
@@ -1,0 +1,219 @@
+# MIT License
+#
+# Copyright (c) 2020 Archis Joglekar
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
+import numpy as np
+
+
+def get_full_leapfrog_step(vdfdx, edfdv, field_solve, x, dt, driver_function):
+    def full_leapfrog_ps_step(e, f, t):
+        """
+        Takes a step forward in time for f and e
+
+        Uses leapfrog scheme
+        1 - spatial advection for 0.5 dt
+
+        2a - field solve
+        2b - velocity advection for dt
+
+        3 - spatial advection for 0.5 dt
+
+        :param f: distribution function. (numpy array of shape (nx, nv))
+        :param x: real-space axis (numpy array of shape (nx,))
+        :param kx: real-space wavenumber axis (numpy array of shape (nx,))
+        :param v: velocity axis (numpy array of shape (nv,))
+        :param kv: velocity-space wavenumber axis (numpy array of shape (nv,))
+        :param dv: velocity-axis spacing (single float value)
+        :param t: current time (single float value)
+        :param dt: timestep (single float value)
+        :param e: electric field (numpy array of shape (nx,))
+        :param driver_function: function that returns an electric field (numpy array of shape (nx,))
+        :return:
+        """
+        # f = vlasov.update_velocity_adv_spectral(f, kv, e, 0.5 * dt)
+        # f = vlasov.update_spatial_adv_spectral(f, kx, v, dt)
+        # e = field.get_total_electric_field(
+        #     driver_function(x, t + dt), f=f, dv=dv, one_over_kx=one_over_kx
+        # )
+        # f = vlasov.update_velocity_adv_spectral(f, kv, e, 0.5 * dt)
+
+        f = edfdv(f=f, e=e, dt=0.5 * dt)
+        f = vdfdx(f=f, dt=dt)
+        e = field_solve(driver_field=driver_function(t + dt), f=f)
+        f = edfdv(f=f, e=e, dt=0.5 * dt)
+
+        return e, f
+
+    return full_leapfrog_ps_step
+
+
+def get_full_pefrl_step(vdfdx, edfdv, field_solve, x, kx, v, kv, dt, driver_function):
+    def full_pefrl_ps_step(e, f, t):
+        """
+        Takes a step forward in time for f and e using the
[…]
