### 2020-08-16 728e85d
passing tests
working health plots

notebooks to go
