### 2020-08-16 180e4fc
added switch for collisionless runs
tested dg and pefrl
