### 2020-08-16 7768650
added pefrl to testing
added moments of f to storage
