### 2020-08-16 84904e8
generalized distribution function storage and added base methods to plotting
still to do notebooks
