### 2020-08-16 ef0add5
added support for looping over different integrators
