### 2020-08-16 a0a4551
finished up landau damping notebook
