+++ tests/test_collisions.py
@@ -21,6 +21,7 @@
 # SOFTWARE.
 
 from itertools import product
+from copy import deepcopy
 
 from tests import helpers
 from vlapy.core import step
@@ -30,12 +31,14 @@ import numpy as np
 ALL_SOLVERS = ["naive", "batched_tridiagonal"]
 ALL_OPERATORS = ["lb", "dg"]
 
+TOLERANCE = 4
+
 
 def __initialize_for_collisions__(vshift):
-    nx = 1
-    nv = 256
+    nx = 2
+    nv = 1024
 
-    nu = 1e-3
+    nu = 1e-2
     dt = 0.1
     v0 = 1.0
 
@@ -65,16 +68,15 @@ def __run_collision_operator_test_loop__(
 
     all_params = {
         "fokker-planck": {"type": collision_operator, "solver": solver},
-        "nu": 1e-3,
+        "nu": nu,
     }
 
     fp_step = step.get_collision_step(
         all_params=all_params, stuff_for_time_loop=stuff_for_time_loop
     )
 
-    f_out = f.copy()
+    f_out = deepcopy(f)
     for it in range(t_end):
-
         f_out = fp_step(f_out)
 
     return f, f_out, v, dv
@@ -96,7 +98,7 @@ def __test_maxwellian_solution__(collision_operator, solver):
         vshift=0.0, t_end=32, collision_operator=collision_operator, solver=solver
     )
 
-    np.testing.assert_almost_equal(f, f_out, decimal=6)
+    np.testing.assert_almost_equal(f, f_out, decimal=4)
 
 
 def test_energy_conservation():
@@ -114,9 +116,9 @@ def __test_energy_conservation__(collision_operator, solver):
         vshift=0.5, t_end=32, collision_operator=collision_operator, solver=solver
     )
 
-    temp_in = np.trapz(f[None,] * v ** 2.0, dx=dv, axis=1)
-    temp_out = np.trapz(f_out[None,] * v ** 2.0, dx=dv, axis=1)
-    np.testing.assert_almost_equal(temp_out, temp_in, decimal=6)
+    temp_in = np.trapz(f * v[None, :] ** 2.0, dx=dv, axis=1)
+    temp_out = np.trapz(f_out * v[None, :] ** 2.0, dx=dv, axis=1)
+    np.testing.assert_almost_equal(temp_out, temp_in, decimal=4)
 
 
 def test_density_conservation():
@@ -135,12 +137,12 @@ def __test_density_conservation__(collision_operator, solver):
         vshift=0.5, t_end=32, collision_operator=collision_operator, solver=solver
     )
 
-    temp_in = np.trapz(f[None,], dx=dv, axis=1)
-    temp_out = np.trapz(f_out[None,], dx=dv, axis=1)
-    np.testing.assert_almost_equal(temp_out, temp_in, decimal=6)
+    temp_in = np.trapz(f, dx=dv, axis=1)
+    temp_out = np.trapz(f_out, dx=dv, axis=1)
+    np.testing.assert_almost_equal(temp_out, temp_in, decimal=TOLERANCE)
 
 
-def test_lenard_bernstein_momentum_conservation_if_initialized_at_zero():
+def test_momentum_conservation_if_initialized_at_zero():
     for collision_operator, solver in product(ALL_OPERATORS, ALL_SOLVERS):
         __test_momentum_conservation_if_initialized_at_zero__(
             collision_operator=collision_operator, solver=solver
@@ -149,21 +151,21 @@ def test_lenard_bernstein_momentum_conservation_if_initialized_at_zero():
 
 def __test_momentum_conservation_if_initialized_at_zero__(collision_operator, solver):
     """
-    tests if the 0th moment of f is conserved
+    tests if the 1st moment of f is conserved if initialized at 0
 
[…]
