### 2020-08-16 144969f
notebooks are working
new plots
