### 2020-08-16 6ccc3df
not testing the naive solver on circle ci because its too slow
