### 2020-08-16 3aaaa75
this was actually fixed when the initializer was modified to create a `np.float64` object. Using this to make a note and pass in a working test in the `run_debug.py` file
