### 2020-08-17 3c0ff56
moved derived quantities to post-loop
helps with processing the driver energy input

still not sure how to exactly account for it, though.
