+++ vlapy/core/vlasov.py
@@ -77,12 +77,29 @@ def get_edfdv_exponential(kv):
 
 def get_edfdv_center_differenced(dv):
     def step_edfdv_center_difference(f, e, dt):
+        """
+        This method calculates the f + dt * e * df/dv using naive
+        2nd-order center differencing
+
+        :param f:
+        :param e:
+        :param dt:
+        :return:
+        """
         return f - e[:, None] * np.gradient(f, dv, axis=1, edge_order=2) * dt
 
     return step_edfdv_center_difference
 
 
 def get_vdfdx(stuff_for_time_loop, vdfdx_implementation="exponential"):
+    """
+    This function enables VlaPy to choose the implementation of the vdfdx stepper
+    to use in the lower level sections of the simulation
+
+    :param stuff_for_time_loop:
+    :param vdfdx_implementation:
+    :return:
+    """
     if vdfdx_implementation == "exponential":
         vdfdx = get_vdfdx_exponential(
             kx=stuff_for_time_loop["kx"], v=stuff_for_time_loop["v"]
@@ -94,6 +111,14 @@ def get_vdfdx(stuff_for_time_loop, vdfdx_implementation="exponential"):
 
 
 def get_edfdv(stuff_for_time_loop, edfdv_implementation="exponential"):
+    """
+    This function enables VlaPy to choose the implementation of the edfdv stepper
+    to use in the lower level sections of the simulation
+
+    :param stuff_for_time_loop:
+    :param edfdv_implementation:
+    :return:
+    """
     if edfdv_implementation == "exponential":
         edfdv = get_edfdv_exponential(kv=stuff_for_time_loop["kv"])
     elif edfdv_implementation == "cd2":
