### 2020-08-17 7d2651e
more docstrings
