+++ vlapy/core/vlasov/linearized_weno5.py
@@ -23,63 +23,247 @@
 import numpy as np
 
 EPSILON = np.float64(1e-6)
+gamma = np.array([0.1, 0.6, 0.3])
 
 
-def get_beta_plus_method():
-    def get_beta_plus_0(f):
-        first_term = np.zeros_like(f)
-        second_term = np.zeros_like(f)
-
-        first_term[:, 0] = 13.0 / 12.0 * (f[:, 0]) ** 2.0
-        first_term[:, 1] = 13.0 / 12.0 * (-2.0 * f[:, 0] + f[:, 1]) ** 2.0
-        first_term[:, 2:] = (
-            13.0 / 12.0 * (f[:, :-2] - 2.0 * f[:, 1:-1] + f[:, 2:]) ** 2.0
+def get_beta_plus_matrices(nv):
+    def get_beta_plus_0_matrices():
+        first_term = (
+            np.diag(np.ones(nv), k=0)
+            + np.diag(-2.0 * np.ones(nv), k=-1)
+            + np.diag(np.ones(nv), k=-2)
+        )
+        second_term = (
+            np.diag(3.0 * np.ones(nv), k=0)
+            + np.diag(-4.0 * np.ones(nv), k=-1)
+            + np.diag(np.ones(nv), k=-2)
         )
 
-        second_term[:, 0] = 1.0 / 4.0 * (3.0 * f[:, 0]) ** 2.0
-        second_term[:, 1] = 1.0 / 4.0 * (-4.0 * f[:, 0] + 3.0 * f[:, 1]) ** 2.0
-        second_term[:, 2:] = (
-            1.0 / 4.0 * (f[:, :-2] - 4.0 * f[:, 1:-1] + 3.0 * f[:, 2:]) ** 2.0
+        return first_term, second_term
+
+    def get_beta_plus_1_matrices():
+        first_term = (
+            np.diag(np.ones(nv), k=1)
+            + np.diag(-2.0 * np.ones(nv), k=0)
+            + np.diag(np.ones(nv), k=-1)
         )
-        return first_term + second_term
+        second_term = np.diag(np.ones(nv), k=1) + np.diag(np.ones(nv), k=-1)
 
-    def get_beta_plus_1(f):
+        return first_term, second_term
 
-        first_term = np.zeros_like(f)
-        second_term = np.zeros_like(f)
-
-        first_term[:, 0] = 13.0 / 12.0 * (-2.0 * f[:, 0] + f[:, 1]) ** 2.0
-        first_term[:, 1:-1] = (
-            13.0 / 12.0 * (f[:, :-2] - 2.0 * f[:, 1:-1] + f[:, 2:]) ** 2.0
+    def get_beta_plus_2_matrices():
+        first_term = (
+            np.diag(np.ones(nv), k=0)
+            + np.diag(-2.0 * np.ones(nv), k=1)
+            + np.diag(np.ones(nv), k=2)
         )
-        first_term[:, -1] = 13.0 / 12.0 * (f[:, -2] - 2.0 * f[:, -1]) ** 2.0
-
-        second_term[:, 0] = 1.0 / 4.0 * (-f[:, 1]) ** 2.0
-        second_term[:, 1:-1] = 1.0 / 4.0 * (f[:, :-2] - f[:, 2:]) ** 2.0
-        second_term[:, -1] = 1.0 / 4.0 * (f[:, -2]) ** 2.0
-
-        return first_term + second_term
-
-    def get_beta_plus_2(f):
-        first_term[:, :-2] = (
-            13.0 / 12.0 * (f[:, :-2] - 2.0 * f[:, 1:-1] + f[:, 2:]) ** 2.0
-        )
[…]
