+++ vlapy/core/vlasov/linearized_weno5.py
@@ -30,37 +30,37 @@ def get_beta_plus_matrices(nv):
     def get_beta_plus_0_matrices():
         first_term = (
             np.diag(np.ones(nv), k=0)
-            + np.diag(-2.0 * np.ones(nv), k=-1)
-            + np.diag(np.ones(nv), k=-2)
+            + np.diag(-2.0 * np.ones(nv - 1), k=-1)
+            + np.diag(np.ones(nv - 2), k=-2)
         )
         second_term = (
             np.diag(3.0 * np.ones(nv), k=0)
-            + np.diag(-4.0 * np.ones(nv), k=-1)
-            + np.diag(np.ones(nv), k=-2)
+            + np.diag(-4.0 * np.ones(nv - 1), k=-1)
+            + np.diag(np.ones(nv - 2), k=-2)
         )
 
         return first_term, second_term
 
     def get_beta_plus_1_matrices():
         first_term = (
-            np.diag(np.ones(nv), k=1)
+            np.diag(np.ones(nv - 1), k=1)
             + np.diag(-2.0 * np.ones(nv), k=0)
-            + np.diag(np.ones(nv), k=-1)
+            + np.diag(np.ones(nv - 1), k=-1)
         )
-        second_term = np.diag(np.ones(nv), k=1) + np.diag(np.ones(nv), k=-1)
+        second_term = np.diag(np.ones(nv - 1), k=1) + np.diag(np.ones(nv - 1), k=-1)
 
         return first_term, second_term
 
     def get_beta_plus_2_matrices():
         first_term = (
             np.diag(np.ones(nv), k=0)
-            + np.diag(-2.0 * np.ones(nv), k=1)
-            + np.diag(np.ones(nv), k=2)
+            + np.diag(-2.0 * np.ones(nv - 1), k=1)
+            + np.diag(np.ones(nv - 2), k=2)
         )
         second_term = (
             np.diag(3.0 * np.ones(nv), k=0)
-            + np.diag(-4.0 * np.ones(nv), k=1)
-            + np.diag(np.ones(nv), k=2)
+            + np.diag(-4.0 * np.ones(nv - 1), k=1)
+            + np.diag(np.ones(nv - 2), k=2)
         )
 
         return first_term, second_term
@@ -75,14 +75,16 @@ def get_beta_plus_matrices(nv):
 def get_beta_minus_matrices(nv):
     def get_beta_minus_0_matrices():
         first_term = (
-            np.diag(np.ones(nv), k=1)
-            + np.diag(-2.0 * np.ones(nv), k=2)
-            + np.diag(np.ones(nv), k=3)
+            np.diag(np.ones(nv - 1), k=1)
+            + np.diag(-2.0 * np.ones(nv - 2), k=2)
+            + np.diag(np.ones(nv - 3), k=3)
+            + np.diag(np.zeros(nv), k=0)
         )
         second_term = (
-            np.diag(3.0 * np.ones(nv), k=1)
-            + np.diag(-4.0 * np.ones(nv), k=2)
-            + np.diag(np.ones(nv), k=3)
+            np.diag(3.0 * np.ones(nv - 1), k=1)
+            + np.diag(-4.0 * np.ones(nv - 2), k=2)
+            + np.diag(np.ones(nv - 3), k=3)
+            + np.diag(np.zeros(nv), k=0)
         )
 
         return first_term, second_term
@@ -90,23 +92,23 @@ def get_beta_minus_matrices(nv):
     def get_beta_minus_1_matrices():
         first_term = (
             np.diag(np.ones(nv), k=0)
-            + np.diag(-2.0 * np.ones(nv), k=1)
[…]
