### 2020-08-18 1d27e1e
into the if statement for the time integrator
