### 2020-08-29 6c3f922
updated readme with screenshots
