### 2020-08-30 5e030b0
more verbose error reports and updated the run files with the new pulse dictionary
