+++ vlapy/core/collisions.py
@@ -24,17 +24,34 @@ import numpy as np
 
 
 def get_philharmonic_matrix_maker(vax, nv, nx, nu, dt, dv):
+    """
+    This function returns the function for preparing the matrix representing the Lenard-Bernstein [1] collision operator.
+
+    It uses the arguments to create static arrays for the linear operato.
+
+    [1] : Lenard, A., & Bernstein, I. B. (1958). Plasma oscillations with diffusion in velocity space.
+        Physical Review, 112(5), 1456–1459. https://doi.org/10.1103/PhysRev.112.1456
+
+    :param vax: (1D float array) - 1D array representing the centers of the velocity grid
+    :param nv: (int) the size of the velocity grid
+    :param nx: (int) the size of the spatial grid
+    :param nu: (float) the collision frequency
+    :param dt: (float) the timestep
+    :param dv: (float) the velocity grid spacing
+    :return: new function with above arguments initialized as static variables
+    """
+
     def make_philharmonic_arrays_for_matrix(f_xv):
         """
-        These are for use with the vectorized collision operator
+        This function creates the arrays representing the Lenard-Bernstein [1] collision operator. When
+        center-differenced to second order in velocity space, this operator can be represented by a tridiagonal matrix.
+        The result of this function is used in a tridiagonal matrix solver.
 
-        :param v: velocity axis (numpy array of shape (nv,))
-        :param nv: size of velocity axis (int)
-        :param nu: collision frequency (float)
-        :param dt: timestep (float)
-        :param dv: velocity-axis spacing (float)
-        :param v0: thermal temperature (float)
-        :return leftside: matrix for the linear operator
+        [1] : Lenard, A., & Bernstein, I. B. (1958). Plasma oscillations with diffusion in velocity space.
+        Physical Review, 112(5), 1456–1459. https://doi.org/10.1103/PhysRev.112.1456
+
+        :param f_xv: The distribution function used to calculate the quantities in the linear operator.
+        :return: a, b, c -- The subdiagonal, diagonal, and super-diagonal, respectively representing the LB operator.
         """
 
         v0t_sq = np.trapz(
@@ -67,18 +84,34 @@ def get_philharmonic_matrix_maker(vax, nv, nx, nu, dt, dv):
 
 
 def get_dougherty_matrix_maker(vax, nv, nx, nu, dt, dv):
+    """
+    This function returns the function for preparing the matrix representing the Dougherty [1] collision operator.
+
+    It uses the arguments to create static arrays for the linear operato.
+
+    [1] : Dougherty, J. P. (1964). Model Fokker-Planck Equation for a Plasma and Its Solution.
+    Physics of Fluids, 7(11), 1788. https://doi.org/10.1063/1.2746779
+
+    :param vax: (1D float array) - 1D array representing the centers of the velocity grid
+    :param nv: (int) the size of the velocity grid
+    :param nx: (int) the size of the spatial grid
+    :param nu: (float) the collision frequency
+    :param dt: (float) the timestep
+    :param dv: (float) the velocity grid spacing
+    :return: new function with above arguments initialized as static variables
+    """
+
     def make_dougherty_arrays_for_matrix(f_xv):
         """
-        This matrix is composed of the linear operator that must be inverted with respect
-        to the right side, which'll be the distribution function
+            This function creates the arrays representing the Dougherty [1] collision operator. When
+            center-differenced to second order in velocity space, this operator can be represented by a tridiagonal matrix.
+            The result of this function is used in a tridiagonal matrix solver.
 
-        :param v: velocity axis (numpy array of shape (nv,))
-        :param nv: size of velocity axis (int)
-        :param nu: collision frequency (float)
[…]
