### 2020-08-31 78e8f2d
collisions is commented
docs updated
