+++ vlapy/core/field.py
@@ -29,8 +29,8 @@ def compute_charges(f, dv):
     Computes a simple moment of the distribution function along
     the velocity axis using the trapezoidal rule
 
-    :param f: distribution function. (numpy array of shape (nx, nv))
-    :param dv: velocity-axis spacing (float)
+    :param f: (2D float array) (nx, nv) the distribution function
+    :param dv: (float) velocity-axis spacing
     :return:
     """
     return np.trapz(f, dx=dv, axis=1)
@@ -41,8 +41,8 @@ def __fft_solve__(net_charge_density, one_over_kx):
     del^2 phi = -rho
     del e = - integral[rho] = - integral[fdv]
 
-    :param net_charge_density: charge-density (numpy array of shape (nx,))
-    :param one_over_kx: one over real-space wavenumber axis (numpy array of shape (nx,))
+    :param net_charge_density: (1D float array (nx,)) charge-density
+    :param one_over_kx: (1D float array (nx,)) one over real-space wavenumber axis (numpy array of shape (nx,))
     :return:
     """
 
@@ -53,8 +53,8 @@ def solve_for_field(charge_density, one_over_kx):
     """
     Solves for the net electric field after subtracting ion charge
 
-    :param charge_density: charge-density (numpy array of shape (nx,))
-    :param one_over_kx: one over real-space wavenumber axis (numpy array of shape (nx,))
+    :param charge_density: (1D float array (nx,)) charge-density
+    :param one_over_kx: (1D float array (nx,)) one over real-space wavenumber axis
 
     :return:
     """
@@ -65,11 +65,12 @@ def solve_for_field(charge_density, one_over_kx):
 
 def get_spectral_solver(dv, one_over_kx):
     """
-    This function gets the field solver
+    This function gets the spectral field solver that uses Fourier transforms to solve the
+    periodic system
 
-    :param dv:
-    :param one_over_kx:
-    :return:
+    :param dv: (float) grid spacing in v
+    :param one_over_kx: (1D float array (nx,)) one over real-space wavenumber axis
+    :return: the function with the above arguments initialized as static variables
     """
 
     def solve_total_electric_field(driver_field, f):
@@ -78,9 +79,7 @@ def get_spectral_solver(dv, one_over_kx):
 
         :param driver_field: an electric field (numpy array of shape (nx,))
         :param f: distribution function. (numpy array of shape (nx, nv))
-        :param dv: velocity-axis spacing (float)
-        :param one_over_kx: one over real-space wavenumber axis (numpy array of shape (nx,))
-        :return:
+        :return: The solver function
         """
         return driver_field + solve_for_field(
             charge_density=compute_charges(f, dv), one_over_kx=one_over_kx
@@ -90,6 +89,13 @@ def get_spectral_solver(dv, one_over_kx):
 
 
 def get_field_solver(stuff_for_time_loop, field_solver_implementation="spectral"):
+    """
+    This method gets the field solver based on the choice in the input parameters.
+
+    :param stuff_for_time_loop: dictionary of parameters for the simulation
+    :param field_solver_implementation:  (string) the name of the field solver chosen in the input parameters
+    :return:
+    """
 
     if field_solver_implementation == "spectral":
         field_solver = get_spectral_solver(
