### 2020-08-31 7154b4c
halfway through storage step
