+++ vlapy/infrastructure/mlflow_helpers.py
@@ -22,11 +22,6 @@
 
 from mlflow.tracking import MlflowClient
 
-
-def get_mlflow_run(experiment_id, run_id):
-    return None
-
-
 def get_this_metric_of_this_run(metric_name, run_object):
     client = MlflowClient()
     run_id = run_object.info.run_id
