### 2020-09-01 7b88080
removed the fftfreq
