### 2020-09-04 0805ced
storing f at every batch
