### 2020-09-05 a9ed8e2
added most of the skeleton code for resume
