+++ vlapy/manager.py
@@ -56,36 +56,16 @@ def start_run(all_params, pulse_dictionary, diagnostics, uris, name="test"):
     with mlflow.start_run(experiment_id=exp_id) as run:
         with tempfile.TemporaryDirectory() as temp_path:
 
-            if diagnostics.rules_to_store_f["space"] == "all":
-                x_storage_multiplier = all_params["nx"]
-            elif diagnostics.rules_to_store_f["space"][0] == "k0":
-                x_storage_multiplier = 2 * len(diagnostics.rules_to_store_f["space"])
-
-            mem_f_store = x_storage_multiplier * all_params["nv"]
-            mem_field_store = all_params["nx"] * 8
-
-            # Initialize loop parameters
-            steps_in_loop = int(
-                1e9
-                * all_params["backend"]["max_GB_for_device"]
-                / (6 * (mem_f_store + mem_field_store) * 8)
+            all_params = get_steps_in_loop_according_to_memory(
+                all_params, diagnostics.rules_to_store_f
             )
-            if steps_in_loop > all_params["nt"]:
-                steps_in_loop = int(all_params["nt"] / 1.25)
-
-            n_loops = all_params["nt"] // steps_in_loop + 1
-            actual_num_steps = n_loops * steps_in_loop
-
-            all_params["steps_in_loop"] = steps_in_loop
-            all_params["n_loops"] = n_loops
-            all_params["actual_num_steps"] = actual_num_steps
 
             # Get numpy arrays of the simulation configuration
             stuff_for_time_loop = outer_loop.get_everything_ready_for_outer_loop(
                 diagnostics=diagnostics,
                 all_params=all_params,
                 pulse_dictionary=pulse_dictionary,
-                overall_num_steps=actual_num_steps,
+                overall_num_steps=all_params["actual_num_steps"],
             )
 
             # Initialize the storage manager -- folders, parameters, etc.
@@ -95,7 +75,7 @@ def start_run(all_params, pulse_dictionary, diagnostics, uris, name="test"):
                 f=stuff_for_time_loop["f"],
                 base_path=temp_path,
                 rules_to_store_f=diagnostics.rules_to_store_f,
-                num_steps_in_one_loop=steps_in_loop,
+                num_steps_in_one_loop=all_params["steps_in_loop"],
                 all_params=all_params,
                 pulse_dictionary=pulse_dictionary,
             )
@@ -106,7 +86,7 @@ def start_run(all_params, pulse_dictionary, diagnostics, uris, name="test"):
             sim_config, do_inner_loop = outer_loop.get_sim_config_and_inner_loop_step(
                 all_params=all_params,
                 stuff_for_time_loop=stuff_for_time_loop,
-                nt_in_loop=steps_in_loop,
+                nt_in_loop=all_params["steps_in_loop"],
                 store_f_rules=diagnostics.rules_to_store_f,
             )
 
@@ -114,7 +94,16 @@ def start_run(all_params, pulse_dictionary, diagnostics, uris, name="test"):
             t0 = time()
 
             # TODO: Could support resume here
-            it_start = 0
+            if all_params["resume"]["bool"]:
+                sim_config, it_start = outer_loop.resume_from_step(
+                    temp_dir=temp_path,
+                    experiment_name=all_params["resume"]["experiment_name"],
+                    run_id=all_params["resume"]["run_id"],
+                    sim_config=sim_config,
+                    resume_time=all_params["resume"]["t0"],
+                )
+            else:
+                it_start = 0
 
             # We run a higher level loop and a lower level loop.
             # The higher level loop contains:
@@ -125,8 +114,14 @@ def start_run(all_params, pulse_dictionary, diagnostics, uris, name="test"):
[…]
