### 2020-09-05 221febd
errors w/ semi lagrangian pusher if timestep is too large
default timestep = 0.05 now
