+++ vlapy/infrastructure/mlflow_helpers.py
@@ -31,10 +31,6 @@ def get_this_metric_of_this_run(metric_name, run_object):
     return run.data.metrics[metric_name]
 
 
-
 def download_run_artifacts_for_resume(path, experiment_name, run_id):
-
-
-
 
     pass
