### 2020-09-05 35841a9
collisions bool
