### 2020-09-06 304ea25
e is recalculated at the beginning of each VP step now
this enables easier resume because only f needs to be loaded

copying from a local mlflow server works fine now, S3 interop still to do

fixing a bug with finding the right time index to restart the run
