+++ resume_vlapy.py
@@ -52,6 +52,10 @@ if __name__ == "__main__":
 
     all_params_dict["fokker-planck"]["bool"] = False
 
+    all_params_dict["resume"]["bool"] = True
+    all_params_dict["resume"]["run_id"] = "0dcc4f66baaf4b62bac521a3413c4402"
+    all_params_dict["resume"]["t0"] = 100
+
     pulse_dictionary = {
         "first pulse": {
             "start_time": 0,
@@ -85,4 +89,3 @@ if __name__ == "__main__":
         uris=uris,
         name=mlflow_exp_name,
     )
-
