### 2020-09-06 834757f
still a bug with resuming
the electric field jumps but the time indexing seems right

needs a bit more of a careful look
