+++ vlapy/storage.py
@@ -29,15 +29,13 @@ import xarray as xr
 import numpy as np
 
 
-def load_f_for_resume(resume_dir, resume_time):
+def load_f_for_resume(f_path, resume_time):
     """
 
     :param resume_dir: (string) path to the folder that contains the distribution function to use to resume
     :return: (2d float array (nx, nv)) distribution function to use for resume
     """
-    overall_path = os.path.join(
-        resume_dir, "full_distribution", "all-full_distribution.nc"
-    )
+    overall_path = os.path.join(f_path, "all-full_distribution.nc")
 
     arr = xr.open_dataset(
         overall_path,
@@ -51,17 +49,17 @@ def load_f_for_resume(resume_dir, resume_time):
         it_start = (
             time_gt_resume_time - 1 + 1
         )  # -1 to get the last stored index, +1 because it_start = 0 isn't stored
-        closest_tax = time_where_f_is_stored[it_start-1]
+        closest_tax = time_where_f_is_stored[it_start - 1]
 
         print("Starting simulation from t = " + str(closest_tax))
+
+        f = arr["full_distribution"].loc[{"time": closest_tax}].data
+
+        return f, it_start
     else:
         raise LookupError(
             "Specified time for resuming simulation is larger than any stored time-steps from the simulation"
         )
-
-    f = arr["full_distribution"].loc[{"time": closest_tax}].data
-
-    return f, it_start
 
 
 def combine_and_save_dataset(individual_path, overall_path):
