### 2020-09-12 #2182 @joglekara
Hi @dpsanders , hope all is well.

I see that you're in the midst of a super cool class this fall so just wanted to help out with a friendly `bump` here ;)
