### 2020-09-12 965df14
Minor edits to paper
