+++ paper/paper.md
@@ -183,7 +183,7 @@ Below, we also illustrate a manual validation of this phenomenon through the ful
 To run the entire testing suite, make sure `pytest` is installed, and call `pytest` from the root folder for the repository. Individual files can also be run by calling `pytest tests/<test_filename>.py`.
 
 # Acknowledgements
-We use xarray [@Hoyer2017] for file storage and MLFlow [@Zaharia2018] for experiment management.
+We use xarray [@Hoyer2017] for file storage and MLFlow [@Chen2020] for experiment management.
 
 We acknowledge valuable discussions with Pierre Navarro on the implementation of the Vlasov equation.
 
