+++ paper/paper.bib
@@ -197,15 +197,20 @@ volume = {5},
 year = {2017}
 }
 
-@article{Zaharia2018,
-  title={Accelerating the Machine Learning Lifecycle with MLflow.},
-  author={Zaharia, Matei and Chen, Andrew and Davidson, Aaron and Ghodsi, Ali and Hong, Sue Ann and Konwinski, Andy and Murching, Siddharth and Nykodym, Tomas and Ogilvie, Paul and Parkhe, Mani and others},
-  journal={IEEE Data Eng. Bull.},
-  volume={41},
-  number={4},
-  pages={39--45},
-  year={2018}
+@inproceedings{Chen2020,
+abstract = {MLflow is a popular open source platform for managing ML development, including experiment tracking, reproducibility, and deployment. In this paper, we discuss user feedback collected since MLflow was launched in 2018, as well as three major features we have introduced in response to this feedback: a Model Registry for collaborative model management and review, tools for simplifying ML code instrumentation, and experiment analytics functions for extracting insights from millions of ML experiments.},
+address = {New York, NY, USA},
+author = {Chen, Andrew and Chow, Andy and Davidson, Aaron and DCunha, Arjun and Ghodsi, Ali and Hong, Sue Ann and Konwinski, Andy and Mewald, Clemens and Murching, Siddharth and Nykodym, Tomas and Ogilvie, Paul and Parkhe, Mani and Singh, Avesh and Xie, Fen and Zaharia, Matei and Zang, Richard and Zheng, Juntai and Zumar, Corey},
+booktitle = {Proceedings of the Fourth International Workshop on Data Management for End-to-End Machine Learning},
+doi = {10.1145/3399579.3399867},
+isbn = {9781450380232},
+publisher = {Association for Computing Machinery},
+series = {DEEM'20},
+title = {{Developments in MLflow: A System to Accelerate the Machine Learning Lifecycle}},
+url = {https://doi.org/10.1145/3399579.3399867},
+year = {2020}
 }
+
 
 @article{Fahlen2009,
 author = {Fahlen, J. E. and Winjum, B. J. and Grismayer, T. and Mori, W. B.},
