### 2020-09-12 #2182 @joglekara
> (By the way, it would be neat to see a version of VlaPy in Julia!).

Yes! That's why I did a huge refactor recently where you can swap the internal solvers.

Next up is pyjulia!
