### 2020-09-12 e03c93c
fully integrated workflow elaboration
