### 2020-09-12 aa33aee
added code to the paper
also snuck in a refactor for the startup messages
