+++ run_nlepw.py
@@ -23,7 +23,6 @@
 import numpy as np
 
 from vlapy import manager, initializers
-from vlapy.infrastructure import mlflow_helpers, print_to_screen
 from vlapy.diagnostics import nlepw
 
 if __name__ == "__main__":
@@ -47,13 +46,6 @@ if __name__ == "__main__":
     all_params_dict["vlasov-poisson"]["edfdv"] = "exponential"
     all_params_dict["vlasov-poisson"]["vdfdx"] = "exponential"
 
-    all_params_dict["backend"]["core"] = "numpy"
-    all_params_dict["backend"]["max_GB_for_device"] = 0.25
-
-    all_params_dict["fokker-planck"]["type"] = "lb"
-    all_params_dict["fokker-planck"]["solver"] = "batched_tridiagonal"
-
-
     pulse_dictionary = {
         "first pulse": {
             "start_time": 0,
@@ -73,10 +65,6 @@ if __name__ == "__main__":
         "tracking": "local",
     }
 
-    print_to_screen.print_startup_message(
-        mlflow_exp_name, all_params_dict, pulse_dictionary
-    )
-
     that_run = manager.start_run(
         all_params=all_params_dict,
         pulse_dictionary=pulse_dictionary,
@@ -87,4 +75,3 @@ if __name__ == "__main__":
         uris=uris,
         name=mlflow_exp_name,
     )
-
