+++ setup.py
@@ -36,7 +36,7 @@ INSTALL_REQUIREMENTS = [
 
 setup(
     name="vlapy",
-    version="1.0",
+    version="0.1",
     packages=find_packages(),
     url="",
     license="MIT",
