### 2020-09-13 #2182 @joglekara
Okay, tagged a release.

Version Number: `VlaPy v0.1`
DOI: `10.5281/zenodo.4026770`
