### 2020-09-16 #2182 @dpsanders
The author list on Zenodo should match the author list in the paper.
