### 2020-09-16 #2182 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
