### 2020-09-17 #2182 @kyleniemeyer
Hi @joglekara, just doing some final checks on the submission.

In the paper:
- In the first sentence, change "1-" to "one-" in both places
- in the second paragraph, please change `(see [@Betti2016; ...` to something like just `[@Betti2016; ...`, or `(see @Betti2016; ...`. The example reference command variants are given at https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography (currently the appearance is a little awkward). Similarly, please change the reference format similarly in the first paragraph in Statement of Need. Personally, I do not think you need "see" or "e.g." in either case, and could just include the lists of references.
- In Equations, please change "4" to "four"
- Just after that, you have a period before "where" (after the normalized variables" that should be a comma
- In the Poisson Equation section, change "1" to "one"
- in the PlasmaPy reference in your .bib file, I think you need curly braces around "PlasmaPy Community"  in the author names list (like `{PlasmaPy Community}`) to avoid BiBTeX treating it like a person's name
