### 2020-09-17 4978428
Cleanup as suggested by JOSS reviewers
