+++ paper/paper.bib
@@ -117,7 +117,7 @@ year = {2002}
 }
 
 @misc{plasmapy,
-  author       = {PlasmaPy Community and
+  author       = {{PlasmaPy Community} and
                   Murphy, Nicholas A. and
                   Leonard, Andrew J. and
                   Stańczak, Dominik and
