### 2020-09-17 1fb6486
numpy and scipy citations
