### 2020-09-17 #2182 @joglekara
Hi @kyleniemeyer , thanks for your eagle-eyed help here! Totally agreed on the 2nd point.

I have addressed all of the changes in the PR above. Let me know if there's anything else I can help with!
