### 2020-09-17 #2182 @joglekara
I should also note that I added citations to NumPy and SciPy (reminded by and in honor of today's publication in Nature) :)
