### 2020-09-17 #2182 @kyleniemeyer
Thanks @joglekara! There were still a few extraneous parentheses left, which I just fixed in https://github.com/joglekara/VlaPy/pull/95
