### 2020-09-17 #2182 @kyleniemeyer
@joglekara thanks for fixing that. It looks like the remaining issue is fine—I can confirm that the DOI for the NumPy matches the one given in the article (probably just isn't through the system yet).
