### 2020-09-17 #2182 @joglekara
Thanks @kyleniemeyer for pushing it through!

Thanks @dpsanders for shepherding the process and to @dpsanders , @TomGoffrey , and @StanczakDominik for the feedback on the submission. I am really impressed by and happy to be contributing to JOSS.
